@@ -17,18 +17,20 @@
  */
 package com.baidu.paddle;
 
+import android.Manifest;
 import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
-import android.graphics.Canvas;
 import android.net.Uri;
-import android.os.AsyncTask;
+import android.os.Build;
 import android.os.Bundle;
 import android.os.Environment;
 import android.provider.MediaStore;
+import android.support.annotation.NonNull;
+import android.support.v4.content.FileProvider;
 import android.util.Log;
 import android.view.View;
 import android.widget.Button;
@@ -36,26 +38,30 @@ import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.baidu.paddle.modeloader.GoogleNetModelCombinedLoaderImpl;
-import com.baidu.paddle.modeloader.GoogleNetModelLoaderCombinedQualifiedImpl;
 import com.baidu.paddle.modeloader.MobileNetModelLoaderImpl;
-import com.baidu.paddle.modeloader.MobileNetModelLoaderQualifiedImpl;
 import com.baidu.paddle.modeloader.ModelLoader;
 
-import java.io.BufferedWriter;
 import java.io.File;
 import java.io.FileOutputStream;
-import java.io.FileWriter;
 import java.io.IOException;
 import java.io.InputStream;
 
+import io.reactivex.Observable;
+import io.reactivex.ObservableEmitter;
+import io.reactivex.Observer;
+import io.reactivex.android.schedulers.AndroidSchedulers;
+import io.reactivex.disposables.Disposable;
+import io.reactivex.functions.Consumer;
+import io.reactivex.schedulers.Schedulers;
 
+
+@SuppressWarnings("ResultOfMethodCallIgnored")
 public class MainActivity extends Activity {
     public static final int TAKE_PHOTO_REQUEST_CODE = 1001;
+    public static final int PERMISSION_REQUEST_CODE = 1002;
 
-    private Context mContext = null;
-
-    private ModelLoader loader = new GoogleNetModelCombinedLoaderImpl();
+    boolean isloaded = false;
+    private ModelLoader loader = new MobileNetModelLoaderImpl();
     ;
 
     private Button btnBanana;
@@ -83,101 +89,131 @@ public class MainActivity extends Activity {
 
     }
 
+    private Uri mOriginUri;
+    private TextView infos;
+    private TextView predictInfos;
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        mContext = this;
         setContentView(R.layout.main_activity);
         init();
+
+        if (!isGotNeededPermissions()) {
+            doRequestPermission();
+        } else {
+            copyModels();
+        }
+
+    }
+
+    private void doRequestPermission() {
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+            String[] permissions = {Manifest.permission.CAMERA, Manifest.permission.WRITE_EXTERNAL_STORAGE};
+            this.requestPermissions(permissions, PERMISSION_REQUEST_CODE);
+        }
     }
 
     private void init() {
-        imageView = (ImageView) findViewById(R.id.imageView);
-        tvSpeed = (TextView) findViewById(R.id.tv_speed);
-        button = (Button) findViewById(R.id.button);
-        btnBanana = (Button) findViewById(R.id.predict_banada);
-        btnBanana.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                scaleImageAndPredictImage();
+        imageView = findViewById(R.id.imageView);
+        tvSpeed = findViewById(R.id.tv_speed);
+        button = findViewById(R.id.button);
+        btnBanana = findViewById(R.id.predict_banada);
+        infos = findViewById(R.id.tv_infos);
+        predictInfos = findViewById(R.id.tv_preinfos);
+        btnBanana.setOnClickListener(view -> scaleImageAndPredictImage(getBanana().getPath()));
+        button.setOnClickListener(view -> {
+            if (!isHasSdCard()) {
+                Toast.makeText(MainActivity.this, R.string.sdcard_not_available,
+                        Toast.LENGTH_LONG).show();
+                return;
             }
-        });
-        button.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                if (!isHasSdCard()) {
-                    Toast.makeText(mContext, R.string.sdcard_not_available,
-                            Toast.LENGTH_LONG).show();
-                    return;
-                }
-                Intent intent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
-                // save pic in sdcard
-                Uri imageUri = Uri.fromFile(getTempImage());
-                intent.putExtra(MediaStore.EXTRA_OUTPUT, imageUri);
-                startActivityForResult(intent, TAKE_PHOTO_REQUEST_CODE);
+//                Intent intent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
+//                // save pic in sdcard
+//                Uri imageUri = Uri.fromFile(getTempImage());
+//                intent.putExtra(MediaStore.EXTRA_OUTPUT, imageUri);
+//                startActivityForResult(intent, TAKE_PHOTO_REQUEST_CODE);
 
-            }
+            takePicFromCamera();
+
         });
         Button bt_load = (Button) findViewById(R.id.bt_load);
-        bt_load.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                loader.load();
-            }
+        bt_load.setOnClickListener(view -> {
+            isloaded = true;
+            loader.load();
         });
         Button bt_clear = (Button) findViewById(R.id.bt_clear);
         bt_clear.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
+                isloaded = false;
+
                 loader.clear();
             }
         });
-        String assetPath = "pml_demo";
-        String sdcardPath = Environment.getExternalStorageDirectory()
-                + File.separator + assetPath;
-        copyFilesFromAssets(this, assetPath, sdcardPath);
 
 
     }
 
-    /**
-     * 缩放然后predict这张图片
-     */
-    @SuppressLint("SetTextI18n")
-    private void scaleImageAndPredictImage() {
-        Bitmap scaleBitmap = getScaleBitmap(
-                MainActivity.this,
-                getBanana().getPath()
-        );
+    private void takePicFromCamera() {
+        Intent intent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
+        intent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
 
-        imageView.setImageBitmap(scaleBitmap);
-        float[] result = loader.predictImage(
-                scaleBitmap
-        );
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+            mOriginUri = FileProvider.getUriForFile(MainActivity.this.getApplication(), MainActivity.this.getApplication().getPackageName() + ".FileProvider",
+                    new File(getTempImage().getPath()));
+        } else {
+            mOriginUri = Uri.fromFile(new File(getTempImage().getPath()));
+        }
+        intent.putExtra(MediaStore.EXTRA_OUTPUT, mOriginUri);
 
-        float max = Float.MIN_VALUE;
-        int maxi = -1;
+        MainActivity.this.startActivityForResult(intent, TAKE_PHOTO_REQUEST_CODE);
+    }
 
-        float sum = 0;
-        if (result != null) {
-            Log.d("pml", "result.length: " + result.length);
+    @SuppressLint("CheckResult")
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
 
-            for (int i = 0; i < result.length; ++i) {
-                Log.d("pml: ", " index: " + i + " value: " + result[i]);
-                sum += result[i];
-                if (result[i] > max) {
-                    max = result[i];
-                    maxi = i;
-                }
+        if (requestCode == PERMISSION_REQUEST_CODE) {
+//            String[] permissions = {Manifest.permission.CAMERA, Manifest.permission.WRITE_EXTERNAL_STORAGE};
+            if (isGotNeededPermissions()) {
+                copyModels();
+            } else {
+                doRequestPermission();
             }
         }
+    }
 
-        Log.d("pml", "maxindex: " + maxi);
-        Log.d("pml", "max: " + max);
-        Log.d("pml", "sum: " + sum);
-        tvSpeed.setText("detection cost：" + loader.getPredictImageTime() + "ms");
-        Toast.makeText(mContext, "maxindex: " + maxi + " max: " + max, Toast.LENGTH_SHORT).show();
+    @SuppressLint({"CheckResult", "SetTextI18n"})
+    private void copyModels() {
+        infos.setText("拷贝模型中....");
+
+        Observable.create((ObservableEmitter<String> emitter) -> {
+            String assetPath = "pml_demo";
+            String sdcardPath = Environment.getExternalStorageDirectory()
+                    + File.separator + assetPath;
+            copyFilesFromAssets(MainActivity.this, assetPath, sdcardPath);
+            emitter.onNext(sdcardPath);
+
+        })
+                .subscribeOn(Schedulers.io())
+                .observeOn(AndroidSchedulers.mainThread(), true)
+                .subscribe(path -> {
+                            infos.setText("模型已拷贝至" + path);
+                            Toast.makeText(MainActivity.this, "模型已拷贝至" + path, Toast.LENGTH_SHORT).show();
+                        }
+                );
+
+
+    }
+
+    private boolean isGotNeededPermissions() {
+        return PermissionUtils.INSTANCE.checkPermissions(this, Manifest.permission.CAMERA)
+                &&
+                PermissionUtils.INSTANCE.checkPermissions(this, Manifest.permission.WRITE_EXTERNAL_STORAGE);
     }
 
     private File getBanana() {
@@ -330,8 +366,11 @@ public class MainActivity extends Activity {
         switch (requestCode) {
             case TAKE_PHOTO_REQUEST_CODE:
                 if (resultCode == RESULT_OK) {
-                    DetectionTask detectionTask = new DetectionTask();
-                    detectionTask.execute(getTempImage().getPath());
+//                    DetectionTask detectionTask = new DetectionTask();
+//                    detectionTask.execute(getTempImage().getPath());
+
+                    scaleImageAndPredictImage(getTempImage().getPath());
+
                 }
                 break;
             default:
@@ -339,73 +378,73 @@ public class MainActivity extends Activity {
         }
     }
 
-    /**
-     * draw rect on imageView
-     *
-     * @param bitmap
-     * @param predicted
-     * @param viewWidth
-     * @param viewHeight
-     */
-    private void drawRect(Bitmap bitmap, float[] predicted, int viewWidth, int viewHeight) {
+//    /**
+//     * draw rect on imageView
+//     *
+//     * @param bitmap
+//     * @param predicted
+//     * @param viewWidth
+//     * @param viewHeight
+//     */
+//    private void drawRect(Bitmap bitmap, float[] predicted, int viewWidth, int viewHeight) {
+//
+//        Canvas canvas = new Canvas(bitmap);
+//        canvas.drawBitmap(bitmap, 0, 0, null);
+//
+//        loader.drawRect(canvas, predicted, viewWidth, viewHeight);
+//
+//        imageView.setImageBitmap(bitmap);
+//
+//    }
+//
+//    private void drawImage(Bitmap bitmap) {
+//        Canvas canvas = new Canvas(bitmap);
+//        canvas.drawBitmap(bitmap, 0, 0, null);
+//        imageView.setImageBitmap(bitmap);
+//    }
+//
+//    float getMaxIndex(float[] predicted) {
+//        float max = 0;
+//        int index = 0;
+//        for (int i = 0; i < predicted.length; i++) {
+//            if (predicted[i] > max) {
+//                max = predicted[i];
+//                index = i;
+//            }
+//        }
+//        return index;
+//    }
 
-        Canvas canvas = new Canvas(bitmap);
-        canvas.drawBitmap(bitmap, 0, 0, null);
-
-        loader.drawRect(canvas, predicted, viewWidth, viewHeight);
-
-        imageView.setImageBitmap(bitmap);
-
-    }
-
-    private void drawImage(Bitmap bitmap) {
-        Canvas canvas = new Canvas(bitmap);
-        canvas.drawBitmap(bitmap, 0, 0, null);
-        imageView.setImageBitmap(bitmap);
-    }
-
-    float getMaxIndex(float[] predicted) {
-        float max = 0;
-        int index = 0;
-        for (int i = 0; i < predicted.length; i++) {
-            if (predicted[i] > max) {
-                max = predicted[i];
-                index = i;
-            }
-        }
-        return index;
-    }
-
-    public float[] getScaledMatrix(Bitmap bitmap, int desWidth,
-                                   int desHeight) {
-        float[] dataBuf = new float[3 * desWidth * desHeight];
-        int rIndex;
-        int gIndex;
-        int bIndex;
-        int[] pixels = new int[desWidth * desHeight];
-        Bitmap bm = Bitmap.createScaledBitmap(bitmap, desWidth, desHeight, false);
-        bm.getPixels(pixels, 0, desWidth, 0, 0, desWidth, desHeight);
-        int j = 0;
-        int k = 0;
-        for (int i = 0; i < pixels.length; i++) {
-            int clr = pixels[i];
-            j = i / desHeight;
-            k = i % desWidth;
-            rIndex = j * desWidth + k;
-            gIndex = rIndex + desHeight * desWidth;
-            bIndex = gIndex + desHeight * desWidth;
-            dataBuf[rIndex] = (float) ((clr & 0x00ff0000) >> 16) - 148;
-            dataBuf[gIndex] = (float) ((clr & 0x0000ff00) >> 8) - 148;
-            dataBuf[bIndex] = (float) ((clr & 0x000000ff)) - 148;
-
-        }
-        if (bm.isRecycled()) {
-            bm.recycle();
-        }
-        return dataBuf;
-
-
-    }
+//    public float[] getScaledMatrix(Bitmap bitmap, int desWidth,
+//                                   int desHeight) {
+//        float[] dataBuf = new float[3 * desWidth * desHeight];
+//        int rIndex;
+//        int gIndex;
+//        int bIndex;
+//        int[] pixels = new int[desWidth * desHeight];
+//        Bitmap bm = Bitmap.createScaledBitmap(bitmap, desWidth, desHeight, false);
+//        bm.getPixels(pixels, 0, desWidth, 0, 0, desWidth, desHeight);
+//        int j = 0;
+//        int k = 0;
+//        for (int i = 0; i < pixels.length; i++) {
+//            int clr = pixels[i];
+//            j = i / desHeight;
+//            k = i % desWidth;
+//            rIndex = j * desWidth + k;
+//            gIndex = rIndex + desHeight * desWidth;
+//            bIndex = gIndex + desHeight * desWidth;
+//            dataBuf[rIndex] = (float) ((clr & 0x00ff0000) >> 16) - 148;
+//            dataBuf[gIndex] = (float) ((clr & 0x0000ff00) >> 8) - 148;
+//            dataBuf[bIndex] = (float) ((clr & 0x000000ff)) - 148;
+//
+//        }
+//        if (bm.isRecycled()) {
+//            bm.recycle();
+//        }
+//        return dataBuf;
+//
+//
+//    }
 
     /**
      * check whether sdcard is mounted
@@ -421,88 +460,139 @@ public class MainActivity extends Activity {
         }
     }
 
-    public void dumpData(float[] results, String filename) {
-        try {
-            File writename = new File(filename);
-            writename.createNewFile();
-            BufferedWriter out = new BufferedWriter(new FileWriter(writename));
-            for (float result : results) {
-                out.write(result + " ");
-            }
-            out.flush();
-            out.close();
-        } catch (IOException e) {
-            e.printStackTrace();
+    /**
+     * 缩放然后predict这张图片
+     */
+    @SuppressLint("SetTextI18n")
+    private void scaleImageAndPredictImage(String path) {
+        if (!isloaded) {
+            loader.load();
         }
+        if (path == null) {
+            Toast.makeText(this, "", Toast.LENGTH_SHORT).show();
+        }
+        Bitmap scaleBitmap = getScaleBitmap(
+                MainActivity.this,
+                path
+        );
+
+        imageView.setImageBitmap(scaleBitmap);
+        float[] result = loader.predictImage(
+                scaleBitmap
+        );
+
+        float max = Float.MIN_VALUE;
+        int maxi = -1;
+
+        float sum = 0;
+        if (result != null) {
+            Log.d("pml", "result.length: " + result.length);
+
+            for (int i = 0; i < result.length; ++i) {
+                Log.d("pml: ", " index: " + i + " value: " + result[i]);
+                sum += result[i];
+                if (result[i] > max) {
+                    max = result[i];
+                    maxi = i;
+                }
+            }
+        }
+
+        Log.d("pml", "maxindex: " + maxi);
+        Log.d("pml", "max: " + max);
+        Log.d("pml", "sum: " + sum);
+        tvSpeed.setText("detection cost：" + loader.getPredictImageTime() + "ms");
+
+        predictInfos.setText("index: " + maxi + " " +
+                "\n结果是: " + MobileNetClassfiedData.INSTANCE.getDataList().get(maxi) +
+                "\n耗时:" + loader.getPredictImageTime() + "ms");
+
+        // Toast.makeText(this, "maxindex: " + maxi + " max: " + max, Toast.LENGTH_SHORT).show();
     }
+//    public void dumpData(float[] results, String filename) {
+//        try {
+//            File writename = new File(filename);
+//            writename.createNewFile();
+//            BufferedWriter out = new BufferedWriter(new FileWriter(writename));
+//            for (float result : results) {
+//                out.write(result + " ");
+//            }
+//            out.flush();
+//            out.close();
+//        } catch (IOException e) {
+//            e.printStackTrace();
+//        }
+//    }
 
     @Override
     public void onBackPressed() {
         super.onBackPressed();
         Log.d("pml", "pml clear");
         // clear pml
+        isloaded = false;
+
         loader.clear();
 
     }
 
-    class DetectionTask extends AsyncTask<String, Void, float[]> {
-        private long time;
-
-        public DetectionTask() {
-            super();
-        }
-
-        @Override
-        protected void onPreExecute() {
-            super.onPreExecute();
-
-
-        }
-
-        @Override
-        protected float[] doInBackground(String... strings) {
-            bmp = getScaleBitmap(mContext, strings[0]);
-
-            float[] result = null;
-            try {
-                result = loader.predictImage(bmp);
-                time = loader.getPredictImageTime();
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-
-            if (result != null) {
-                for (int i = 0; i < result.length; ++i) {
-                    Log.d("pml: ", " index: " + i + " value: " + result[i]);
-                }
-            }
-
-            return result;
-        }
-
-        @Override
-        protected void onPostExecute(float[] result) {
-            super.onPostExecute(result);
-            try {
-                Bitmap src = Bitmap.createScaledBitmap(bmp, imageView.getWidth(),
-                        imageView.getHeight(), false);
-                drawRect(src, result, imageView.getWidth(), imageView.getHeight());
-                tvSpeed.setText("detection cost：" + time + "ms");
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-        }
-
-        @Override
-        protected void onProgressUpdate(Void... values) {
-            super.onProgressUpdate(values);
-        }
-
-        @Override
-        protected void onCancelled() {
-            super.onCancelled();
-        }
-
-
-    }
+//    class DetectionTask extends AsyncTask<String, Void, float[]> {
+//        private long time;
+//
+//        public DetectionTask() {
+//            super();
+//        }
+//
+//        @Override
+//        protected void onPreExecute() {
+//            super.onPreExecute();
+//
+//
+//        }
+//
+//        @Override
+//        protected float[] doInBackground(String... strings) {
+//            bmp = getScaleBitmap(mContext, strings[0]);
+//
+//            float[] result = null;
+//            try {
+//                result = loader.predictImage(bmp);
+//                time = loader.getPredictImageTime();
+//            } catch (Exception e) {
+//                e.printStackTrace();
+//            }
+//
+//            if (result != null) {
+//                for (int i = 0; i < result.length; ++i) {
+//                    Log.d("pml: ", " index: " + i + " value: " + result[i]);
+//                }
+//            }
+//
+//            return result;
+//        }
+//
+//        @Override
+//        protected void onPostExecute(float[] result) {
+//            super.onPostExecute(result);
+//            try {
+//                Bitmap src = Bitmap.createScaledBitmap(bmp, imageView.getWidth(),
+//                        imageView.getHeight(), false);
+//                drawRect(src, result, imageView.getWidth(), imageView.getHeight());
+//                tvSpeed.setText("detection cost：" + time + "ms");
+//            } catch (Exception e) {
+//                e.printStackTrace();
+//            }
+//        }
+//
+//        @Override
+//        protected void onProgressUpdate(Void... values) {
+//            super.onProgressUpdate(values);
+//        }
+//
+//        @Override
+//        protected void onCancelled() {
+//            super.onCancelled();
+//        }
+//
+//
+//    }
 }
