@@ -3,1006 +3,1006 @@ package com.baidu.paddle
 object MobileNetClassfiedData {
     val dataList by lazy{
         val al = ArrayList<String>()
-        al.add("n01440764 tench, Tinca tinca")
-        al.add("n01443537 goldfish, Carassius auratus")
-        al.add("n01484850 great white shark, white shark, man-eater, man-eating shark, Carcharodon carcharias")
-        al.add("n01491361 tiger shark, Galeocerdo cuvieri")
-        al.add("n01494475 hammerhead, hammerhead shark")
-        al.add("n01496331 electric ray, crampfish, numbfish, torpedo")
-        al.add("n01498041 stingray")
-        al.add("n01514668 cock")
-        al.add("n01514859 hen")
-        al.add("n01518878 ostrich, Struthio camelus")
-        al.add("n01530575 brambling, Fringilla montifringilla")
-        al.add("n01531178 goldfinch, Carduelis carduelis")
-        al.add("n01532829 house finch, linnet, Carpodacus mexicanus")
-        al.add("n01534433 junco, snowbird")
-        al.add("n01537544 indigo bunting, indigo finch, indigo bird, Passerina cyanea")
-        al.add("n01558993 robin, American robin, Turdus migratorius")
-        al.add("n01560419 bulbul")
-        al.add("n01580077 jay")
-        al.add("n01582220 magpie")
-        al.add("n01592084 chickadee")
-        al.add("n01601694 water ouzel, dipper")
-        al.add("n01608432 kite")
-        al.add("n01614925 bald eagle, American eagle, Haliaeetus leucocephalus")
-        al.add("n01616318 vulture")
-        al.add("n01622779 great grey owl, great gray owl, Strix nebulosa")
-        al.add("n01629819 European fire salamander, Salamandra salamandra")
-        al.add("n01630670 common newt, Triturus vulgaris")
-        al.add("n01631663 eft")
-        al.add("n01632458 spotted salamander, Ambystoma maculatum")
-        al.add("n01632777 axolotl, mud puppy, Ambystoma mexicanum")
-        al.add("n01641577 bullfrog, Rana catesbeiana")
-        al.add("n01644373 tree frog, tree-frog")
-        al.add("n01644900 tailed frog, bell toad, ribbed toad, tailed toad, Ascaphus trui")
-        al.add("n01664065 loggerhead, loggerhead turtle, Caretta caretta")
-        al.add("n01665541 leatherback turtle, leatherback, leathery turtle, Dermochelys coriacea")
-        al.add("n01667114 mud turtle")
-        al.add("n01667778 terrapin")
-        al.add("n01669191 box turtle, box tortoise")
-        al.add("n01675722 banded gecko")
-        al.add("n01677366 common iguana, iguana, Iguana iguana")
-        al.add("n01682714 American chameleon, anole, Anolis carolinensis")
-        al.add("n01685808 whiptail, whiptail lizard")
-        al.add("n01687978 agama")
-        al.add("n01688243 frilled lizard, Chlamydosaurus kingi")
-        al.add("n01689811 alligator lizard")
-        al.add("n01692333 Gila monster, Heloderma suspectum")
-        al.add("n01693334 green lizard, Lacerta viridis")
-        al.add("n01694178 African chameleon, Chamaeleo chamaeleon")
-        al.add("n01695060 Komodo dragon, Komodo lizard, dragon lizard, giant lizard, Varanus komodoensis")
-        al.add("n01697457 African crocodile, Nile crocodile, Crocodylus niloticus")
-        al.add("n01698640 American alligator, Alligator mississipiensis")
-        al.add("n01704323 triceratops")
-        al.add("n01728572 thunder snake, worm snake, Carphophis amoenus")
-        al.add("n01728920 ringneck snake, ring-necked snake, ring snake")
-        al.add("n01729322 hognose snake, puff adder, sand viper")
-        al.add("n01729977 green snake, grass snake")
-        al.add("n01734418 king snake, kingsnake")
-        al.add("n01735189 garter snake, grass snake")
-        al.add("n01737021 water snake")
-        al.add("n01739381 vine snake")
-        al.add("n01740131 night snake, Hypsiglena torquata")
-        al.add("n01742172 boa constrictor, Constrictor constrictor")
-        al.add("n01744401 rock python, rock snake, Python sebae")
-        al.add("n01748264 Indian cobra, Naja naja")
-        al.add("n01749939 green mamba")
-        al.add("n01751748 sea snake")
-        al.add("n01753488 horned viper, cerastes, sand viper, horned asp, Cerastes cornutus")
-        al.add("n01755581 diamondback, diamondback rattlesnake, Crotalus adamanteus")
-        al.add("n01756291 sidewinder, horned rattlesnake, Crotalus cerastes")
-        al.add("n01768244 trilobite")
-        al.add("n01770081 harvestman, daddy longlegs, Phalangium opilio")
-        al.add("n01770393 scorpion")
-        al.add("n01773157 black and gold garden spider, Argiope aurantia")
-        al.add("n01773549 barn spider, Araneus cavaticus")
-        al.add("n01773797 garden spider, Aranea diademata")
-        al.add("n01774384 black widow, Latrodectus mactans")
-        al.add("n01774750 tarantula")
-        al.add("n01775062 wolf spider, hunting spider")
-        al.add("n01776313 tick")
-        al.add("n01784675 centipede")
-        al.add("n01795545 black grouse")
-        al.add("n01796340 ptarmigan")
-        al.add("n01797886 ruffed grouse, partridge, Bonasa umbellus")
-        al.add("n01798484 prairie chicken, prairie grouse, prairie fowl")
-        al.add("n01806143 peacock")
-        al.add("n01806567 quail")
-        al.add("n01807496 partridge")
-        al.add("n01817953 African grey, African gray, Psittacus erithacus")
-        al.add("n01818515 macaw")
-        al.add("n01819313 sulphur-crested cockatoo, Kakatoe galerita, Cacatua galerita")
-        al.add("n01820546 lorikeet")
-        al.add("n01824575 coucal")
-        al.add("n01828970 bee eater")
-        al.add("n01829413 hornbill")
-        al.add("n01833805 hummingbird")
-        al.add("n01843065 jacamar")
-        al.add("n01843383 toucan")
-        al.add("n01847000 drake")
-        al.add("n01855032 red-breasted merganser, Mergus serrator")
-        al.add("n01855672 goose")
-        al.add("n01860187 black swan, Cygnus atratus")
-        al.add("n01871265 tusker")
-        al.add("n01872401 echidna, spiny anteater, anteater")
-        al.add("n01873310 platypus, duckbill, duckbilled platypus, duck-billed platypus, Ornithorhynchus anatinus")
-        al.add("n01877812 wallaby, brush kangaroo")
-        al.add("n01882714 koala, koala bear, kangaroo bear, native bear, Phascolarctos cinereus")
-        al.add("n01883070 wombat")
-        al.add("n01910747 jellyfish")
-        al.add("n01914609 sea anemone, anemone")
-        al.add("n01917289 brain coral")
-        al.add("n01924916 flatworm, platyhelminth")
-        al.add("n01930112 nematode, nematode worm, roundworm")
-        al.add("n01943899 conch")
-        al.add("n01944390 snail")
-        al.add("n01945685 slug")
-        al.add("n01950731 sea slug, nudibranch")
-        al.add("n01955084 chiton, coat-of-mail shell, sea cradle, polyplacophore")
-        al.add("n01968897 chambered nautilus, pearly nautilus, nautilus")
-        al.add("n01978287 Dungeness crab, Cancer magister")
-        al.add("n01978455 rock crab, Cancer irroratus")
-        al.add("n01980166 fiddler crab")
-        al.add("n01981276 king crab, Alaska crab, Alaskan king crab, Alaska king crab, Paralithodes camtschatica")
-        al.add("n01983481 American lobster, Northern lobster, Maine lobster, Homarus americanus")
-        al.add("n01984695 spiny lobster, langouste, rock lobster, crawfish, crayfish, sea crawfish")
-        al.add("n01985128 crayfish, crawfish, crawdad, crawdaddy")
-        al.add("n01986214 hermit crab")
-        al.add("n01990800 isopod")
-        al.add("n02002556 white stork, Ciconia ciconia")
-        al.add("n02002724 black stork, Ciconia nigra")
-        al.add("n02006656 spoonbill")
-        al.add("n02007558 flamingo")
-        al.add("n02009229 little blue heron, Egretta caerulea")
-        al.add("n02009912 American egret, great white heron, Egretta albus")
-        al.add("n02011460 bittern")
-        al.add("n02012849 crane")
-        al.add("n02013706 limpkin, Aramus pictus")
-        al.add("n02017213 European gallinule, Porphyrio porphyrio")
-        al.add("n02018207 American coot, marsh hen, mud hen, water hen, Fulica americana")
-        al.add("n02018795 bustard")
-        al.add("n02025239 ruddy turnstone, Arenaria interpres")
-        al.add("n02027492 red-backed sandpiper, dunlin, Erolia alpina")
-        al.add("n02028035 redshank, Tringa totanus")
-        al.add("n02033041 dowitcher")
-        al.add("n02037110 oystercatcher, oyster catcher")
-        al.add("n02051845 pelican")
-        al.add("n02056570 king penguin, Aptenodytes patagonica")
-        al.add("n02058221 albatross, mollymawk")
-        al.add("n02066245 grey whale, gray whale, devilfish, Eschrichtius gibbosus, Eschrichtius robustus")
-        al.add("n02071294 killer whale, killer, orca, grampus, sea wolf, Orcinus orca")
-        al.add("n02074367 dugong, Dugong dugon")
-        al.add("n02077923 sea lion")
-        al.add("n02085620 Chihuahua")
-        al.add("n02085782 Japanese spaniel")
-        al.add("n02085936 Maltese dog, Maltese terrier, Maltese")
-        al.add("n02086079 Pekinese, Pekingese, Peke")
-        al.add("n02086240 Shih-Tzu")
-        al.add("n02086646 Blenheim spaniel")
-        al.add("n02086910 papillon")
-        al.add("n02087046 toy terrier")
-        al.add("n02087394 Rhodesian ridgeback")
-        al.add("n02088094 Afghan hound, Afghan")
-        al.add("n02088238 basset, basset hound")
-        al.add("n02088364 beagle")
-        al.add("n02088466 bloodhound, sleuthhound")
-        al.add("n02088632 bluetick")
-        al.add("n02089078 black-and-tan coonhound")
-        al.add("n02089867 Walker hound, Walker foxhound")
-        al.add("n02089973 English foxhound")
-        al.add("n02090379 redbone")
-        al.add("n02090622 borzoi, Russian wolfhound")
-        al.add("n02090721 Irish wolfhound")
-        al.add("n02091032 Italian greyhound")
-        al.add("n02091134 whippet")
-        al.add("n02091244 Ibizan hound, Ibizan Podenco")
-        al.add("n02091467 Norwegian elkhound, elkhound")
-        al.add("n02091635 otterhound, otter hound")
-        al.add("n02091831 Saluki, gazelle hound")
-        al.add("n02092002 Scottish deerhound, deerhound")
-        al.add("n02092339 Weimaraner")
-        al.add("n02093256 Staffordshire bullterrier, Staffordshire bull terrier")
-        al.add("n02093428 American Staffordshire terrier, Staffordshire terrier, American pit bull terrier, pit bull terrier")
-        al.add("n02093647 Bedlington terrier")
-        al.add("n02093754 Border terrier")
-        al.add("n02093859 Kerry blue terrier")
-        al.add("n02093991 Irish terrier")
-        al.add("n02094114 Norfolk terrier")
-        al.add("n02094258 Norwich terrier")
-        al.add("n02094433 Yorkshire terrier")
-        al.add("n02095314 wire-haired fox terrier")
-        al.add("n02095570 Lakeland terrier")
-        al.add("n02095889 Sealyham terrier, Sealyham")
-        al.add("n02096051 Airedale, Airedale terrier")
-        al.add("n02096177 cairn, cairn terrier")
-        al.add("n02096294 Australian terrier")
-        al.add("n02096437 Dandie Dinmont, Dandie Dinmont terrier")
-        al.add("n02096585 Boston bull, Boston terrier")
-        al.add("n02097047 miniature schnauzer")
-        al.add("n02097130 giant schnauzer")
-        al.add("n02097209 standard schnauzer")
-        al.add("n02097298 Scotch terrier, Scottish terrier, Scottie")
-        al.add("n02097474 Tibetan terrier, chrysanthemum dog")
-        al.add("n02097658 silky terrier, Sydney silky")
-        al.add("n02098105 soft-coated wheaten terrier")
-        al.add("n02098286 West Highland white terrier")
-        al.add("n02098413 Lhasa, Lhasa apso")
-        al.add("n02099267 flat-coated retriever")
-        al.add("n02099429 curly-coated retriever")
-        al.add("n02099601 golden retriever")
-        al.add("n02099712 Labrador retriever")
-        al.add("n02099849 Chesapeake Bay retriever")
-        al.add("n02100236 German short-haired pointer")
-        al.add("n02100583 vizsla, Hungarian pointer")
-        al.add("n02100735 English setter")
-        al.add("n02100877 Irish setter, red setter")
-        al.add("n02101006 Gordon setter")
-        al.add("n02101388 Brittany spaniel")
-        al.add("n02101556 clumber, clumber spaniel")
-        al.add("n02102040 English springer, English springer spaniel")
-        al.add("n02102177 Welsh springer spaniel")
-        al.add("n02102318 cocker spaniel, English cocker spaniel, cocker")
-        al.add("n02102480 Sussex spaniel")
-        al.add("n02102973 Irish water spaniel")
-        al.add("n02104029 kuvasz")
-        al.add("n02104365 schipperke")
-        al.add("n02105056 groenendael")
-        al.add("n02105162 malinois")
-        al.add("n02105251 briard")
-        al.add("n02105412 kelpie")
-        al.add("n02105505 komondor")
-        al.add("n02105641 Old English sheepdog, bobtail")
-        al.add("n02105855 Shetland sheepdog, Shetland sheep dog, Shetland")
-        al.add("n02106030 collie")
-        al.add("n02106166 Border collie")
-        al.add("n02106382 Bouvier des Flandres, Bouviers des Flandres")
-        al.add("n02106550 Rottweiler")
-        al.add("n02106662 German shepherd, German shepherd dog, German police dog, alsatian")
-        al.add("n02107142 Doberman, Doberman pinscher")
-        al.add("n02107312 miniature pinscher")
-        al.add("n02107574 Greater Swiss Mountain dog")
-        al.add("n02107683 Bernese mountain dog")
-        al.add("n02107908 Appenzeller")
-        al.add("n02108000 EntleBucher")
-        al.add("n02108089 boxer")
-        al.add("n02108422 bull mastiff")
-        al.add("n02108551 Tibetan mastiff")
-        al.add("n02108915 French bulldog")
-        al.add("n02109047 Great Dane")
-        al.add("n02109525 Saint Bernard, St Bernard")
-        al.add("n02109961 Eskimo dog, husky")
-        al.add("n02110063 malamute, malemute, Alaskan malamute")
-        al.add("n02110185 Siberian husky")
-        al.add("n02110341 dalmatian, coach dog, carriage dog")
-        al.add("n02110627 affenpinscher, monkey pinscher, monkey dog")
-        al.add("n02110806 basenji")
-        al.add("n02110958 pug, pug-dog")
-        al.add("n02111129 Leonberg")
-        al.add("n02111277 Newfoundland, Newfoundland dog")
-        al.add("n02111500 Great Pyrenees")
-        al.add("n02111889 Samoyed, Samoyede")
-        al.add("n02112018 Pomeranian")
-        al.add("n02112137 chow, chow chow")
-        al.add("n02112350 keeshond")
-        al.add("n02112706 Brabancon griffon")
-        al.add("n02113023 Pembroke, Pembroke Welsh corgi")
-        al.add("n02113186 Cardigan, Cardigan Welsh corgi")
-        al.add("n02113624 toy poodle")
-        al.add("n02113712 miniature poodle")
-        al.add("n02113799 standard poodle")
-        al.add("n02113978 Mexican hairless")
-        al.add("n02114367 timber wolf, grey wolf, gray wolf, Canis lupus")
-        al.add("n02114548 white wolf, Arctic wolf, Canis lupus tundrarum")
-        al.add("n02114712 red wolf, maned wolf, Canis rufus, Canis niger")
-        al.add("n02114855 coyote, prairie wolf, brush wolf, Canis latrans")
-        al.add("n02115641 dingo, warrigal, warragal, Canis dingo")
-        al.add("n02115913 dhole, Cuon alpinus")
-        al.add("n02116738 African hunting dog, hyena dog, Cape hunting dog, Lycaon pictus")
-        al.add("n02117135 hyena, hyaena")
-        al.add("n02119022 red fox, Vulpes vulpes")
-        al.add("n02119789 kit fox, Vulpes macrotis")
-        al.add("n02120079 Arctic fox, white fox, Alopex lagopus")
-        al.add("n02120505 grey fox, gray fox, Urocyon cinereoargenteus")
-        al.add("n02123045 tabby, tabby cat")
-        al.add("n02123159 tiger cat")
-        al.add("n02123394 Persian cat")
-        al.add("n02123597 Siamese cat, Siamese")
-        al.add("n02124075 Egyptian cat")
-        al.add("n02125311 cougar, puma, catamount, mountain lion, painter, panther, Felis concolor")
-        al.add("n02127052 lynx, catamount")
-        al.add("n02128385 leopard, Panthera pardus")
-        al.add("n02128757 snow leopard, ounce, Panthera uncia")
-        al.add("n02128925 jaguar, panther, Panthera onca, Felis onca")
-        al.add("n02129165 lion, king of beasts, Panthera leo")
-        al.add("n02129604 tiger, Panthera tigris")
-        al.add("n02130308 cheetah, chetah, Acinonyx jubatus")
-        al.add("n02132136 brown bear, bruin, Ursus arctos")
-        al.add("n02133161 American black bear, black bear, Ursus americanus, Euarctos americanus")
-        al.add("n02134084 ice bear, polar bear, Ursus Maritimus, Thalarctos maritimus")
-        al.add("n02134418 sloth bear, Melursus ursinus, Ursus ursinus")
-        al.add("n02137549 mongoose")
-        al.add("n02138441 meerkat, mierkat")
-        al.add("n02165105 tiger beetle")
-        al.add("n02165456 ladybug, ladybeetle, lady beetle, ladybird, ladybird beetle")
-        al.add("n02167151 ground beetle, carabid beetle")
-        al.add("n02168699 long-horned beetle, longicorn, longicorn beetle")
-        al.add("n02169497 leaf beetle, chrysomelid")
-        al.add("n02172182 dung beetle")
-        al.add("n02174001 rhinoceros beetle")
-        al.add("n02177972 weevil")
-        al.add("n02190166 fly")
-        al.add("n02206856 bee")
-        al.add("n02219486 ant, emmet, pismire")
-        al.add("n02226429 grasshopper, hopper")
-        al.add("n02229544 cricket")
-        al.add("n02231487 walking stick, walkingstick, stick insect")
-        al.add("n02233338 cockroach, roach")
-        al.add("n02236044 mantis, mantid")
-        al.add("n02256656 cicada, cicala")
-        al.add("n02259212 leafhopper")
-        al.add("n02264363 lacewing, lacewing fly")
-        al.add("n02268443 dragonfly, darning needle, devil''s darning needle, sewing needle, snake feeder, snake doctor, mosquito hawk, skeeter hawk")
-        al.add("n02268853 damselfly")
-        al.add("n02276258 admiral")
-        al.add("n02277742 ringlet, ringlet butterfly")
-        al.add("n02279972 monarch, monarch butterfly, milkweed butterfly, Danaus plexippus")
-        al.add("n02280649 cabbage butterfly")
-        al.add("n02281406 sulphur butterfly, sulfur butterfly")
-        al.add("n02281787 lycaenid, lycaenid butterfly")
-        al.add("n02317335 starfish, sea star")
-        al.add("n02319095 sea urchin")
-        al.add("n02321529 sea cucumber, holothurian")
-        al.add("n02325366 wood rabbit, cottontail, cottontail rabbit")
-        al.add("n02326432 hare")
-        al.add("n02328150 Angora, Angora rabbit")
-        al.add("n02342885 hamster")
-        al.add("n02346627 porcupine, hedgehog")
-        al.add("n02356798 fox squirrel, eastern fox squirrel, Sciurus niger")
-        al.add("n02361337 marmot")
-        al.add("n02363005 beaver")
-        al.add("n02364673 guinea pig, Cavia cobaya")
-        al.add("n02389026 sorrel")
-        al.add("n02391049 zebra")
-        al.add("n02395406 hog, pig, grunter, squealer, Sus scrofa")
-        al.add("n02396427 wild boar, boar, Sus scrofa")
-        al.add("n02397096 warthog")
-        al.add("n02398521 hippopotamus, hippo, river horse, Hippopotamus amphibius")
-        al.add("n02403003 ox")
-        al.add("n02408429 water buffalo, water ox, Asiatic buffalo, Bubalus bubalis")
-        al.add("n02410509 bison")
-        al.add("n02412080 ram, tup")
-        al.add("n02415577 bighorn, bighorn sheep, cimarron, Rocky Mountain bighorn, Rocky Mountain sheep, Ovis canadensis")
-        al.add("n02417914 ibex, Capra ibex")
-        al.add("n02422106 hartebeest")
-        al.add("n02422699 impala, Aepyceros melampus")
-        al.add("n02423022 gazelle")
-        al.add("n02437312 Arabian camel, dromedary, Camelus dromedarius")
-        al.add("n02437616 llama")
-        al.add("n02441942 weasel")
-        al.add("n02442845 mink")
-        al.add("n02443114 polecat, fitch, foulmart, foumart, Mustela putorius")
-        al.add("n02443484 black-footed ferret, ferret, Mustela nigripes")
-        al.add("n02444819 otter")
-        al.add("n02445715 skunk, polecat, wood pussy")
-        al.add("n02447366 badger")
-        al.add("n02454379 armadillo")
-        al.add("n02457408 three-toed sloth, ai, Bradypus tridactylus")
-        al.add("n02480495 orangutan, orang, orangutang, Pongo pygmaeus")
-        al.add("n02480855 gorilla, Gorilla gorilla")
-        al.add("n02481823 chimpanzee, chimp, Pan troglodytes")
-        al.add("n02483362 gibbon, Hylobates lar")
-        al.add("n02483708 siamang, Hylobates syndactylus, Symphalangus syndactylus")
-        al.add("n02484975 guenon, guenon monkey")
-        al.add("n02486261 patas, hussar monkey, Erythrocebus patas")
-        al.add("n02486410 baboon")
-        al.add("n02487347 macaque")
-        al.add("n02488291 langur")
-        al.add("n02488702 colobus, colobus monkey")
-        al.add("n02489166 proboscis monkey, Nasalis larvatus")
-        al.add("n02490219 marmoset")
-        al.add("n02492035 capuchin, ringtail, Cebus capucinus")
-        al.add("n02492660 howler monkey, howler")
-        al.add("n02493509 titi, titi monkey")
-        al.add("n02493793 spider monkey, Ateles geoffroyi")
-        al.add("n02494079 squirrel monkey, Saimiri sciureus")
-        al.add("n02497673 Madagascar cat, ring-tailed lemur, Lemur catta")
-        al.add("n02500267 indri, indris, Indri indri, Indri brevicaudatus")
-        al.add("n02504013 Indian elephant, Elephas maximus")
-        al.add("n02504458 African elephant, Loxodonta africana")
-        al.add("n02509815 lesser panda, red panda, panda, bear cat, cat bear, Ailurus fulgens")
-        al.add("n02510455 giant panda, panda, panda bear, coon bear, Ailuropoda melanoleuca")
-        al.add("n02514041 barracouta, snoek")
-        al.add("n02526121 eel")
-        al.add("n02536864 coho, cohoe, coho salmon, blue jack, silver salmon, Oncorhynchus kisutch")
-        al.add("n02606052 rock beauty, Holocanthus tricolor")
-        al.add("n02607072 anemone fish")
-        al.add("n02640242 sturgeon")
-        al.add("n02641379 gar, garfish, garpike, billfish, Lepisosteus osseus")
-        al.add("n02643566 lionfish")
-        al.add("n02655020 puffer, pufferfish, blowfish, globefish")
-        al.add("n02666196 abacus")
-        al.add("n02667093 abaya")
-        al.add("n02669723 academic gown, academic robe, judge''s robe")
-        al.add("n02672831 accordion, piano accordion, squeeze box")
-        al.add("n02676566 acoustic guitar")
-        al.add("n02687172 aircraft carrier, carrier, flattop, attack aircraft carrier")
-        al.add("n02690373 airliner")
-        al.add("n02692877 airship, dirigible")
-        al.add("n02699494 altar")
-        al.add("n02701002 ambulance")
-        al.add("n02704792 amphibian, amphibious vehicle")
-        al.add("n02708093 analog clock")
-        al.add("n02727426 apiary, bee house")
-        al.add("n02730930 apron")
-        al.add("n02747177 ashcan, trash can, garbage can, wastebin, ash bin, ash-bin, ashbin, dustbin, trash barrel, trash bin")
-        al.add("n02749479 assault rifle, assault gun")
-        al.add("n02769748 backpack, back pack, knapsack, packsack, rucksack, haversack")
-        al.add("n02776631 bakery, bakeshop, bakehouse")
-        al.add("n02777292 balance beam, beam")
-        al.add("n02782093 balloon")
-        al.add("n02783161 ballpoint, ballpoint pen, ballpen, Biro")
-        al.add("n02786058 Band Aid")
-        al.add("n02787622 banjo")
-        al.add("n02788148 bannister, banister, balustrade, balusters, handrail")
-        al.add("n02790996 barbell")
-        al.add("n02791124 barber chair")
-        al.add("n02791270 barbershop")
-        al.add("n02793495 barn")
-        al.add("n02794156 barometer")
-        al.add("n02795169 barrel, cask")
-        al.add("n02797295 barrow, garden cart, lawn cart, wheelbarrow")
-        al.add("n02799071 baseball")
-        al.add("n02802426 basketball")
-        al.add("n02804414 bassinet")
-        al.add("n02804610 bassoon")
-        al.add("n02807133 bathing cap, swimming cap")
-        al.add("n02808304 bath towel")
-        al.add("n02808440 bathtub, bathing tub, bath, tub")
-        al.add("n02814533 beach wagon, station wagon, wagon, estate car, beach waggon, station waggon, waggon")
-        al.add("n02814860 beacon, lighthouse, beacon light, pharos")
-        al.add("n02815834 beaker")
-        al.add("n02817516 bearskin, busby, shako")
-        al.add("n02823428 beer bottle")
-        al.add("n02823750 beer glass")
-        al.add("n02825657 bell cote, bell cot")
-        al.add("n02834397 bib")
-        al.add("n02835271 bicycle-built-for-two, tandem bicycle, tandem")
-        al.add("n02837789 bikini, two-piece")
-        al.add("n02840245 binder, ring-binder")
-        al.add("n02841315 binoculars, field glasses, opera glasses")
-        al.add("n02843684 birdhouse")
-        al.add("n02859443 boathouse")
-        al.add("n02860847 bobsled, bobsleigh, bob")
-        al.add("n02865351 bolo tie, bolo, bola tie, bola")
-        al.add("n02869837 bonnet, poke bonnet")
-        al.add("n02870880 bookcase")
-        al.add("n02871525 bookshop, bookstore, bookstall")
-        al.add("n02877765 bottlecap")
-        al.add("n02879718 bow")
-        al.add("n02883205 bow tie, bow-tie, bowtie")
-        al.add("n02892201 brass, memorial tablet, plaque")
-        al.add("n02892767 brassiere, bra, bandeau")
-        al.add("n02894605 breakwater, groin, groyne, mole, bulwark, seawall, jetty")
-        al.add("n02895154 breastplate, aegis, egis")
-        al.add("n02906734 broom")
-        al.add("n02909870 bucket, pail")
-        al.add("n02910353 buckle")
-        al.add("n02916936 bulletproof vest")
-        al.add("n02917067 bullet train, bullet")
-        al.add("n02927161 butcher shop, meat market")
-        al.add("n02930766 cab, hack, taxi, taxicab")
-        al.add("n02939185 caldron, cauldron")
-        al.add("n02948072 candle, taper, wax light")
-        al.add("n02950826 cannon")
-        al.add("n02951358 canoe")
-        al.add("n02951585 can opener, tin opener")
-        al.add("n02963159 cardigan")
-        al.add("n02965783 car mirror")
-        al.add("n02966193 carousel, carrousel, merry-go-round, roundabout, whirligig")
-        al.add("n02966687 carpenter''s kit, tool kit")
-        al.add("n02971356 carton")
-        al.add("n02974003 car wheel")
-        al.add("n02977058 cash machine, cash dispenser, automated teller machine, automatic teller machine, automated teller, automatic teller, ATM")
-        al.add("n02978881 cassette")
-        al.add("n02979186 cassette player")
-        al.add("n02980441 castle")
-        al.add("n02981792 catamaran")
-        al.add("n02988304 CD player")
-        al.add("n02992211 cello, violoncello")
-        al.add("n02992529 cellular telephone, cellular phone, cellphone, cell, mobile phone")
-        al.add("n02999410 chain")
-        al.add("n03000134 chainlink fence")
-        al.add("n03000247 chain mail, ring mail, mail, chain armor, chain armour, ring armor, ring armour")
-        al.add("n03000684 chain saw, chainsaw")
-        al.add("n03014705 chest")
-        al.add("n03016953 chiffonier, commode")
-        al.add("n03017168 chime, bell, gong")
-        al.add("n03018349 china cabinet, china closet")
-        al.add("n03026506 Christmas stocking")
-        al.add("n03028079 church, church building")
-        al.add("n03032252 cinema, movie theater, movie theatre, movie house, picture palace")
-        al.add("n03041632 cleaver, meat cleaver, chopper")
-        al.add("n03042490 cliff dwelling")
-        al.add("n03045698 cloak")
-        al.add("n03047690 clog, geta, patten, sabot")
-        al.add("n03062245 cocktail shaker")
-        al.add("n03063599 coffee mug")
-        al.add("n03063689 coffeepot")
-        al.add("n03065424 coil, spiral, volute, whorl, helix")
-        al.add("n03075370 combination lock")
-        al.add("n03085013 computer keyboard, keypad")
-        al.add("n03089624 confectionery, confectionary, candy store")
-        al.add("n03095699 container ship, containership, container vessel")
-        al.add("n03100240 convertible")
-        al.add("n03109150 corkscrew, bottle screw")
-        al.add("n03110669 cornet, horn, trumpet, trump")
-        al.add("n03124043 cowboy boot")
-        al.add("n03124170 cowboy hat, ten-gallon hat")
-        al.add("n03125729 cradle")
-        al.add("n03126707 crane")
-        al.add("n03127747 crash helmet")
-        al.add("n03127925 crate")
-        al.add("n03131574 crib, cot")
-        al.add("n03133878 Crock Pot")
-        al.add("n03134739 croquet ball")
-        al.add("n03141823 crutch")
-        al.add("n03146219 cuirass")
-        al.add("n03160309 dam, dike, dyke")
-        al.add("n03179701 desk")
-        al.add("n03180011 desktop computer")
-        al.add("n03187595 dial telephone, dial phone")
-        al.add("n03188531 diaper, nappy, napkin")
-        al.add("n03196217 digital clock")
-        al.add("n03197337 digital watch")
-        al.add("n03201208 dining table, board")
-        al.add("n03207743 dishrag, dishcloth")
-        al.add("n03207941 dishwasher, dish washer, dishwashing machine")
-        al.add("n03208938 disk brake, disc brake")
-        al.add("n03216828 dock, dockage, docking facility")
-        al.add("n03218198 dogsled, dog sled, dog sleigh")
-        al.add("n03220513 dome")
-        al.add("n03223299 doormat, welcome mat")
-        al.add("n03240683 drilling platform, offshore rig")
-        al.add("n03249569 drum, membranophone, tympan")
-        al.add("n03250847 drumstick")
-        al.add("n03255030 dumbbell")
-        al.add("n03259280 Dutch oven")
-        al.add("n03271574 electric fan, blower")
-        al.add("n03272010 electric guitar")
-        al.add("n03272562 electric locomotive")
-        al.add("n03290653 entertainment center")
-        al.add("n03291819 envelope")
-        al.add("n03297495 espresso maker")
-        al.add("n03314780 face powder")
-        al.add("n03325584 feather boa, boa")
-        al.add("n03337140 file, file cabinet, filing cabinet")
-        al.add("n03344393 fireboat")
-        al.add("n03345487 fire engine, fire truck")
-        al.add("n03347037 fire screen, fireguard")
-        al.add("n03355925 flagpole, flagstaff")
-        al.add("n03372029 flute, transverse flute")
-        al.add("n03376595 folding chair")
-        al.add("n03379051 football helmet")
-        al.add("n03384352 forklift")
-        al.add("n03388043 fountain")
-        al.add("n03388183 fountain pen")
-        al.add("n03388549 four-poster")
-        al.add("n03393912 freight car")
-        al.add("n03394916 French horn, horn")
-        al.add("n03400231 frying pan, frypan, skillet")
-        al.add("n03404251 fur coat")
-        al.add("n03417042 garbage truck, dustcart")
-        al.add("n03424325 gasmask, respirator, gas helmet")
-        al.add("n03425413 gas pump, gasoline pump, petrol pump, island dispenser")
-        al.add("n03443371 goblet")
-        al.add("n03444034 go-kart")
-        al.add("n03445777 golf ball")
-        al.add("n03445924 golfcart, golf cart")
-        al.add("n03447447 gondola")
-        al.add("n03447721 gong, tam-tam")
-        al.add("n03450230 gown")
-        al.add("n03452741 grand piano, grand")
-        al.add("n03457902 greenhouse, nursery, glasshouse")
-        al.add("n03459775 grille, radiator grille")
-        al.add("n03461385 grocery store, grocery, food market, market")
-        al.add("n03467068 guillotine")
-        al.add("n03476684 hair slide")
-        al.add("n03476991 hair spray")
-        al.add("n03478589 half track")
-        al.add("n03481172 hammer")
-        al.add("n03482405 hamper")
-        al.add("n03483316 hand blower, blow dryer, blow drier, hair dryer, hair drier")
-        al.add("n03485407 hand-held computer, hand-held microcomputer")
-        al.add("n03485794 handkerchief, hankie, hanky, hankey")
-        al.add("n03492542 hard disc, hard disk, fixed disk")
-        al.add("n03494278 harmonica, mouth organ, harp, mouth harp")
-        al.add("n03495258 harp")
-        al.add("n03496892 harvester, reaper")
-        al.add("n03498962 hatchet")
-        al.add("n03527444 holster")
-        al.add("n03529860 home theater, home theatre")
-        al.add("n03530642 honeycomb")
-        al.add("n03532672 hook, claw")
-        al.add("n03534580 hoopskirt, crinoline")
-        al.add("n03535780 horizontal bar, high bar")
-        al.add("n03538406 horse cart, horse-cart")
-        al.add("n03544143 hourglass")
-        al.add("n03584254 iPod")
-        al.add("n03584829 iron, smoothing iron")
-        al.add("n03590841 jack-o''-lantern")
-        al.add("n03594734 jean, blue jean, denim")
-        al.add("n03594945 jeep, landrover")
-        al.add("n03595614 jersey, T-shirt, tee shirt")
-        al.add("n03598930 jigsaw puzzle")
-        al.add("n03599486 jinrikisha, ricksha, rickshaw")
-        al.add("n03602883 joystick")
-        al.add("n03617480 kimono")
-        al.add("n03623198 knee pad")
-        al.add("n03627232 knot")
-        al.add("n03630383 lab coat, laboratory coat")
-        al.add("n03633091 ladle")
-        al.add("n03637318 lampshade, lamp shade")
-        al.add("n03642806 laptop, laptop computer")
-        al.add("n03649909 lawn mower, mower")
-        al.add("n03657121 lens cap, lens cover")
-        al.add("n03658185 letter opener, paper knife, paperknife")
-        al.add("n03661043 library")
-        al.add("n03662601 lifeboat")
-        al.add("n03666591 lighter, light, igniter, ignitor")
-        al.add("n03670208 limousine, limo")
-        al.add("n03673027 liner, ocean liner")
-        al.add("n03676483 lipstick, lip rouge")
-        al.add("n03680355 Loafer")
-        al.add("n03690938 lotion")
-        al.add("n03691459 loudspeaker, speaker, speaker unit, loudspeaker system, speaker system")
-        al.add("n03692522 loupe, jeweler''s loupe")
-        al.add("n03697007 lumbermill, sawmill")
-        al.add("n03706229 magnetic compass")
-        al.add("n03709823 mailbag, postbag")
-        al.add("n03710193 mailbox, letter box")
-        al.add("n03710637 maillot")
-        al.add("n03710721 maillot, tank suit")
-        al.add("n03717622 manhole cover")
-        al.add("n03720891 maraca")
-        al.add("n03721384 marimba, xylophone")
-        al.add("n03724870 mask")
-        al.add("n03729826 matchstick")
-        al.add("n03733131 maypole")
-        al.add("n03733281 maze, labyrinth")
-        al.add("n03733805 measuring cup")
-        al.add("n03742115 medicine chest, medicine cabinet")
-        al.add("n03743016 megalith, megalithic structure")
-        al.add("n03759954 microphone, mike")
-        al.add("n03761084 microwave, microwave oven")
-        al.add("n03763968 military uniform")
-        al.add("n03764736 milk can")
-        al.add("n03769881 minibus")
-        al.add("n03770439 miniskirt, mini")
-        al.add("n03770679 minivan")
-        al.add("n03773504 missile")
-        al.add("n03775071 mitten")
-        al.add("n03775546 mixing bowl")
-        al.add("n03776460 mobile home, manufactured home")
-        al.add("n03777568 Model T")
-        al.add("n03777754 modem")
-        al.add("n03781244 monastery")
-        al.add("n03782006 monitor")
-        al.add("n03785016 moped")
-        al.add("n03786901 mortar")
-        al.add("n03787032 mortarboard")
-        al.add("n03788195 mosque")
-        al.add("n03788365 mosquito net")
-        al.add("n03791053 motor scooter, scooter")
-        al.add("n03792782 mountain bike, all-terrain bike, off-roader")
-        al.add("n03792972 mountain tent")
-        al.add("n03793489 mouse, computer mouse")
-        al.add("n03794056 mousetrap")
-        al.add("n03796401 moving van")
-        al.add("n03803284 muzzle")
-        al.add("n03804744 nail")
-        al.add("n03814639 neck brace")
-        al.add("n03814906 necklace")
-        al.add("n03825788 nipple")
-        al.add("n03832673 notebook, notebook computer")
-        al.add("n03837869 obelisk")
-        al.add("n03838899 oboe, hautboy, hautbois")
-        al.add("n03840681 ocarina, sweet potato")
-        al.add("n03841143 odometer, hodometer, mileometer, milometer")
-        al.add("n03843555 oil filter")
-        al.add("n03854065 organ, pipe organ")
-        al.add("n03857828 oscilloscope, scope, cathode-ray oscilloscope, CRO")
-        al.add("n03866082 overskirt")
-        al.add("n03868242 oxcart")
-        al.add("n03868863 oxygen mask")
-        al.add("n03871628 packet")
-        al.add("n03873416 paddle, boat paddle")
-        al.add("n03874293 paddlewheel, paddle wheel")
-        al.add("n03874599 padlock")
-        al.add("n03876231 paintbrush")
-        al.add("n03877472 pajama, pyjama, pj''s, jammies")
-        al.add("n03877845 palace")
-        al.add("n03884397 panpipe, pandean pipe, syrinx")
-        al.add("n03887697 paper towel")
-        al.add("n03888257 parachute, chute")
-        al.add("n03888605 parallel bars, bars")
-        al.add("n03891251 park bench")
-        al.add("n03891332 parking meter")
-        al.add("n03895866 passenger car, coach, carriage")
-        al.add("n03899768 patio, terrace")
-        al.add("n03902125 pay-phone, pay-station")
-        al.add("n03903868 pedestal, plinth, footstall")
-        al.add("n03908618 pencil box, pencil case")
-        al.add("n03908714 pencil sharpener")
-        al.add("n03916031 perfume, essence")
-        al.add("n03920288 Petri dish")
-        al.add("n03924679 photocopier")
-        al.add("n03929660 pick, plectrum, plectron")
-        al.add("n03929855 pickelhaube")
-        al.add("n03930313 picket fence, paling")
-        al.add("n03930630 pickup, pickup truck")
-        al.add("n03933933 pier")
-        al.add("n03935335 piggy bank, penny bank")
-        al.add("n03937543 pill bottle")
-        al.add("n03938244 pillow")
-        al.add("n03942813 ping-pong ball")
-        al.add("n03944341 pinwheel")
-        al.add("n03947888 pirate, pirate ship")
-        al.add("n03950228 pitcher, ewer")
-        al.add("n03954731 plane, carpenter''s plane, woodworking plane")
-        al.add("n03956157 planetarium")
-        al.add("n03958227 plastic bag")
-        al.add("n03961711 plate rack")
-        al.add("n03967562 plow, plough")
-        al.add("n03970156 plunger, plumber''s helper")
-        al.add("n03976467 Polaroid camera, Polaroid Land camera")
-        al.add("n03976657 pole")
-        al.add("n03977966 police van, police wagon, paddy wagon, patrol wagon, wagon, black Maria")
-        al.add("n03980874 poncho")
-        al.add("n03982430 pool table, billiard table, snooker table")
-        al.add("n03983396 pop bottle, soda bottle")
-        al.add("n03991062 pot, flowerpot")
-        al.add("n03992509 potter''s wheel")
-        al.add("n03995372 power drill")
-        al.add("n03998194 prayer rug, prayer mat")
-        al.add("n04004767 printer")
-        al.add("n04005630 prison, prison house")
-        al.add("n04008634 projectile, missile")
-        al.add("n04009552 projector")
-        al.add("n04019541 puck, hockey puck")
-        al.add("n04023962 punching bag, punch bag, punching ball, punchball")
-        al.add("n04026417 purse")
-        al.add("n04033901 quill, quill pen")
-        al.add("n04033995 quilt, comforter, comfort, puff")
-        al.add("n04037443 racer, race car, racing car")
-        al.add("n04039381 racket, racquet")
-        al.add("n04040759 radiator")
-        al.add("n04041544 radio, wireless")
-        al.add("n04044716 radio telescope, radio reflector")
-        al.add("n04049303 rain barrel")
-        al.add("n04065272 recreational vehicle, RV, R.V.")
-        al.add("n04067472 reel")
-        al.add("n04069434 reflex camera")
-        al.add("n04070727 refrigerator, icebox")
-        al.add("n04074963 remote control, remote")
-        al.add("n04081281 restaurant, eating house, eating place, eatery")
-        al.add("n04086273 revolver, six-gun, six-shooter")
-        al.add("n04090263 rifle")
-        al.add("n04099969 rocking chair, rocker")
-        al.add("n04111531 rotisserie")
-        al.add("n04116512 rubber eraser, rubber, pencil eraser")
-        al.add("n04118538 rugby ball")
-        al.add("n04118776 rule, ruler")
-        al.add("n04120489 running shoe")
-        al.add("n04125021 safe")
-        al.add("n04127249 safety pin")
-        al.add("n04131690 saltshaker, salt shaker")
-        al.add("n04133789 sandal")
-        al.add("n04136333 sarong")
-        al.add("n04141076 sax, saxophone")
-        al.add("n04141327 scabbard")
-        al.add("n04141975 scale, weighing machine")
-        al.add("n04146614 school bus")
-        al.add("n04147183 schooner")
-        al.add("n04149813 scoreboard")
-        al.add("n04152593 screen, CRT screen")
-        al.add("n04153751 screw")
-        al.add("n04154565 screwdriver")
-        al.add("n04162706 seat belt, seatbelt")
-        al.add("n04179913 sewing machine")
-        al.add("n04192698 shield, buckler")
-        al.add("n04200800 shoe shop, shoe-shop, shoe store")
-        al.add("n04201297 shoji")
-        al.add("n04204238 shopping basket")
-        al.add("n04204347 shopping cart")
-        al.add("n04208210 shovel")
-        al.add("n04209133 shower cap")
-        al.add("n04209239 shower curtain")
-        al.add("n04228054 ski")
-        al.add("n04229816 ski mask")
-        al.add("n04235860 sleeping bag")
-        al.add("n04238763 slide rule, slipstick")
-        al.add("n04239074 sliding door")
-        al.add("n04243546 slot, one-armed bandit")
-        al.add("n04251144 snorkel")
-        al.add("n04252077 snowmobile")
-        al.add("n04252225 snowplow, snowplough")
-        al.add("n04254120 soap dispenser")
-        al.add("n04254680 soccer ball")
-        al.add("n04254777 sock")
-        al.add("n04258138 solar dish, solar collector, solar furnace")
-        al.add("n04259630 sombrero")
-        al.add("n04263257 soup bowl")
-        al.add("n04264628 space bar")
-        al.add("n04265275 space heater")
-        al.add("n04266014 space shuttle")
-        al.add("n04270147 spatula")
-        al.add("n04273569 speedboat")
-        al.add("n04275548 spider web, spider''s web")
-        al.add("n04277352 spindle")
-        al.add("n04285008 sports car, sport car")
-        al.add("n04286575 spotlight, spot")
-        al.add("n04296562 stage")
-        al.add("n04310018 steam locomotive")
-        al.add("n04311004 steel arch bridge")
-        al.add("n04311174 steel drum")
-        al.add("n04317175 stethoscope")
-        al.add("n04325704 stole")
-        al.add("n04326547 stone wall")
-        al.add("n04328186 stopwatch, stop watch")
-        al.add("n04330267 stove")
-        al.add("n04332243 strainer")
-        al.add("n04335435 streetcar, tram, tramcar, trolley, trolley car")
-        al.add("n04336792 stretcher")
-        al.add("n04344873 studio couch, day bed")
-        al.add("n04346328 stupa, tope")
-        al.add("n04347754 submarine, pigboat, sub, U-boat")
-        al.add("n04350905 suit, suit of clothes")
-        al.add("n04355338 sundial")
-        al.add("n04355933 sunglass")
-        al.add("n04356056 sunglasses, dark glasses, shades")
-        al.add("n04357314 sunscreen, sunblock, sun blocker")
-        al.add("n04366367 suspension bridge")
-        al.add("n04367480 swab, swob, mop")
-        al.add("n04370456 sweatshirt")
-        al.add("n04371430 swimming trunks, bathing trunks")
-        al.add("n04371774 swing")
-        al.add("n04372370 switch, electric switch, electrical switch")
-        al.add("n04376876 syringe")
-        al.add("n04380533 table lamp")
-        al.add("n04389033 tank, army tank, armored combat vehicle, armoured combat vehicle")
-        al.add("n04392985 tape player")
-        al.add("n04398044 teapot")
-        al.add("n04399382 teddy, teddy bear")
-        al.add("n04404412 television, television system")
-        al.add("n04409515 tennis ball")
-        al.add("n04417672 thatch, thatched roof")
-        al.add("n04418357 theater curtain, theatre curtain")
-        al.add("n04423845 thimble")
-        al.add("n04428191 thresher, thrasher, threshing machine")
-        al.add("n04429376 throne")
-        al.add("n04435653 tile roof")
-        al.add("n04442312 toaster")
-        al.add("n04443257 tobacco shop, tobacconist shop, tobacconist")
-        al.add("n04447861 toilet seat")
-        al.add("n04456115 torch")
-        al.add("n04458633 totem pole")
-        al.add("n04461696 tow truck, tow car, wrecker")
-        al.add("n04462240 toyshop")
-        al.add("n04465501 tractor")
-        al.add("n04467665 trailer truck, tractor trailer, trucking rig, rig, articulated lorry, semi")
-        al.add("n04476259 tray")
-        al.add("n04479046 trench coat")
-        al.add("n04482393 tricycle, trike, velocipede")
-        al.add("n04483307 trimaran")
-        al.add("n04485082 tripod")
-        al.add("n04486054 triumphal arch")
-        al.add("n04487081 trolleybus, trolley coach, trackless trolley")
-        al.add("n04487394 trombone")
-        al.add("n04493381 tub, vat")
-        al.add("n04501370 turnstile")
-        al.add("n04505470 typewriter keyboard")
-        al.add("n04507155 umbrella")
-        al.add("n04509417 unicycle, monocycle")
-        al.add("n04515003 upright, upright piano")
-        al.add("n04517823 vacuum, vacuum cleaner")
-        al.add("n04522168 vase")
-        al.add("n04523525 vault")
-        al.add("n04525038 velvet")
-        al.add("n04525305 vending machine")
-        al.add("n04532106 vestment")
-        al.add("n04532670 viaduct")
-        al.add("n04536866 violin, fiddle")
-        al.add("n04540053 volleyball")
-        al.add("n04542943 waffle iron")
-        al.add("n04548280 wall clock")
-        al.add("n04548362 wallet, billfold, notecase, pocketbook")
-        al.add("n04550184 wardrobe, closet, press")
-        al.add("n04552348 warplane, military plane")
-        al.add("n04553703 washbasin, handbasin, washbowl, lavabo, wash-hand basin")
-        al.add("n04554684 washer, automatic washer, washing machine")
-        al.add("n04557648 water bottle")
-        al.add("n04560804 water jug")
-        al.add("n04562935 water tower")
-        al.add("n04579145 whiskey jug")
-        al.add("n04579432 whistle")
-        al.add("n04584207 wig")
-        al.add("n04589890 window screen")
-        al.add("n04590129 window shade")
-        al.add("n04591157 Windsor tie")
-        al.add("n04591713 wine bottle")
-        al.add("n04592741 wing")
-        al.add("n04596742 wok")
-        al.add("n04597913 wooden spoon")
-        al.add("n04599235 wool, woolen, woollen")
-        al.add("n04604644 worm fence, snake fence, snake-rail fence, Virginia fence")
-        al.add("n04606251 wreck")
-        al.add("n04612504 yawl")
-        al.add("n04613696 yurt")
-        al.add("n06359193 web site, website, internet site, site")
-        al.add("n06596364 comic book")
-        al.add("n06785654 crossword puzzle, crossword")
-        al.add("n06794110 street sign")
-        al.add("n06874185 traffic light, traffic signal, stoplight")
-        al.add("n07248320 book jacket, dust cover, dust jacket, dust wrapper")
-        al.add("n07565083 menu")
-        al.add("n07579787 plate")
-        al.add("n07583066 guacamole")
-        al.add("n07584110 consomme")
-        al.add("n07590611 hot pot, hotpot")
-        al.add("n07613480 trifle")
-        al.add("n07614500 ice cream, icecream")
-        al.add("n07615774 ice lolly, lolly, lollipop, popsicle")
-        al.add("n07684084 French loaf")
-        al.add("n07693725 bagel, beigel")
-        al.add("n07695742 pretzel")
-        al.add("n07697313 cheeseburger")
-        al.add("n07697537 hotdog, hot dog, red hot")
-        al.add("n07711569 mashed potato")
-        al.add("n07714571 head cabbage")
-        al.add("n07714990 broccoli")
-        al.add("n07715103 cauliflower")
-        al.add("n07716358 zucchini, courgette")
-        al.add("n07716906 spaghetti squash")
-        al.add("n07717410 acorn squash")
-        al.add("n07717556 butternut squash")
-        al.add("n07718472 cucumber, cuke")
-        al.add("n07718747 artichoke, globe artichoke")
-        al.add("n07720875 bell pepper")
-        al.add("n07730033 cardoon")
-        al.add("n07734744 mushroom")
-        al.add("n07742313 Granny Smith")
-        al.add("n07745940 strawberry")
-        al.add("n07747607 orange")
-        al.add("n07749582 lemon")
-        al.add("n07753113 fig")
-        al.add("n07753275 pineapple, ananas")
-        al.add("n07753592 banana")
-        al.add("n07754684 jackfruit, jak, jack")
-        al.add("n07760859 custard apple")
-        al.add("n07768694 pomegranate")
-        al.add("n07802026 hay")
-        al.add("n07831146 carbonara")
-        al.add("n07836838 chocolate sauce, chocolate syrup")
-        al.add("n07860988 dough")
-        al.add("n07871810 meat loaf, meatloaf")
-        al.add("n07873807 pizza, pizza pie")
-        al.add("n07875152 potpie")
-        al.add("n07880968 burrito")
-        al.add("n07892512 red wine")
-        al.add("n07920052 espresso")
-        al.add("n07930864 cup")
-        al.add("n07932039 eggnog")
-        al.add("n09193705 alp")
-        al.add("n09229709 bubble")
-        al.add("n09246464 cliff, drop, drop-off")
-        al.add("n09256479 coral reef")
-        al.add("n09288635 geyser")
-        al.add("n09332890 lakeside, lakeshore")
-        al.add("n09399592 promontory, headland, head, foreland")
-        al.add("n09421951 sandbar, sand bar")
-        al.add("n09428293 seashore, coast, seacoast, sea-coast")
-        al.add("n09468604 valley, vale")
-        al.add("n09472597 volcano")
-        al.add("n09835506 ballplayer, baseball player")
-        al.add("n10148035 groom, bridegroom")
-        al.add("n10565667 scuba diver")
-        al.add("n11879895 rapeseed")
-        al.add("n11939491 daisy")
-        al.add("n12057211 yellow lady''s slipper, yellow lady-slipper, Cypripedium calceolus, Cypripedium parviflorum")
-        al.add("n12144580 corn")
-        al.add("n12267677 acorn")
-        al.add("n12620546 hip, rose hip, rosehip")
-        al.add("n12768682 buckeye, horse chestnut, conker")
-        al.add("n12985857 coral fungus")
-        al.add("n12998815 agaric")
-        al.add("n13037406 gyromitra")
-        al.add("n13040303 stinkhorn, carrion fungus")
-        al.add("n13044778 earthstar")
-        al.add("n13052670 hen-of-the-woods, hen of the woods, Polyporus frondosus, Grifola frondosa")
-        al.add("n13054560 bolete")
-        al.add("n13133613 ear, spike, capitulum")
-        al.add("n15075141 toilet tissue, toilet paper, bathroom tissue")
+        al.add("tench, Tinca tinca")
+        al.add("goldfish, Carassius auratus")
+        al.add("great white shark, white shark, man-eater, man-eating shark, Carcharodon carcharias")
+        al.add("tiger shark, Galeocerdo cuvieri")
+        al.add("hammerhead, hammerhead shark")
+        al.add("electric ray, crampfish, numbfish, torpedo")
+        al.add("stingray")
+        al.add("cock")
+        al.add("hen")
+        al.add("ostrich, Struthio camelus")
+        al.add("brambling, Fringilla montifringilla")
+        al.add("goldfinch, Carduelis carduelis")
+        al.add("house finch, linnet, Carpodacus mexicanus")
+        al.add("junco, snowbird")
+        al.add("indigo bunting, indigo finch, indigo bird, Passerina cyanea")
+        al.add("robin, American robin, Turdus migratorius")
+        al.add("bulbul")
+        al.add("jay")
+        al.add("magpie")
+        al.add("chickadee")
+        al.add("water ouzel, dipper")
+        al.add("kite")
+        al.add("bald eagle, American eagle, Haliaeetus leucocephalus")
+        al.add("vulture")
+        al.add("great grey owl, great gray owl, Strix nebulosa")
+        al.add("European fire salamander, Salamandra salamandra")
+        al.add("common newt, Triturus vulgaris")
+        al.add("eft")
+        al.add("spotted salamander, Ambystoma maculatum")
+        al.add("axolotl, mud puppy, Ambystoma mexicanum")
+        al.add("bullfrog, Rana catesbeiana")
+        al.add("tree frog, tree-frog")
+        al.add("tailed frog, bell toad, ribbed toad, tailed toad, Ascaphus trui")
+        al.add("loggerhead, loggerhead turtle, Caretta caretta")
+        al.add("leatherback turtle, leatherback, leathery turtle, Dermochelys coriacea")
+        al.add("mud turtle")
+        al.add("terrapin")
+        al.add("box turtle, box tortoise")
+        al.add("banded gecko")
+        al.add("common iguana, iguana, Iguana iguana")
+        al.add("American chameleon, anole, Anolis carolinensis")
+        al.add("whiptail, whiptail lizard")
+        al.add("agama")
+        al.add("frilled lizard, Chlamydosaurus kingi")
+        al.add("alligator lizard")
+        al.add("Gila monster, Heloderma suspectum")
+        al.add("green lizard, Lacerta viridis")
+        al.add("African chameleon, Chamaeleo chamaeleon")
+        al.add("Komodo dragon, Komodo lizard, dragon lizard, giant lizard, Varanus komodoensis")
+        al.add("African crocodile, Nile crocodile, Crocodylus niloticus")
+        al.add("American alligator, Alligator mississipiensis")
+        al.add("triceratops")
+        al.add("thunder snake, worm snake, Carphophis amoenus")
+        al.add("ringneck snake, ring-necked snake, ring snake")
+        al.add("hognose snake, puff adder, sand viper")
+        al.add("green snake, grass snake")
+        al.add("king snake, kingsnake")
+        al.add("garter snake, grass snake")
+        al.add("water snake")
+        al.add("vine snake")
+        al.add("night snake, Hypsiglena torquata")
+        al.add("boa constrictor, Constrictor constrictor")
+        al.add("rock python, rock snake, Python sebae")
+        al.add("Indian cobra, Naja naja")
+        al.add("green mamba")
+        al.add("sea snake")
+        al.add("horned viper, cerastes, sand viper, horned asp, Cerastes cornutus")
+        al.add("diamondback, diamondback rattlesnake, Crotalus adamanteus")
+        al.add("sidewinder, horned rattlesnake, Crotalus cerastes")
+        al.add("trilobite")
+        al.add("harvestman, daddy longlegs, Phalangium opilio")
+        al.add("scorpion")
+        al.add("black and gold garden spider, Argiope aurantia")
+        al.add("barn spider, Araneus cavaticus")
+        al.add("garden spider, Aranea diademata")
+        al.add("black widow, Latrodectus mactans")
+        al.add("tarantula")
+        al.add("wolf spider, hunting spider")
+        al.add("tick")
+        al.add("centipede")
+        al.add("black grouse")
+        al.add("ptarmigan")
+        al.add("ruffed grouse, partridge, Bonasa umbellus")
+        al.add("prairie chicken, prairie grouse, prairie fowl")
+        al.add("peacock")
+        al.add("quail")
+        al.add("partridge")
+        al.add("African grey, African gray, Psittacus erithacus")
+        al.add("macaw")
+        al.add("sulphur-crested cockatoo, Kakatoe galerita, Cacatua galerita")
+        al.add("lorikeet")
+        al.add("coucal")
+        al.add("bee eater")
+        al.add("hornbill")
+        al.add("hummingbird")
+        al.add("jacamar")
+        al.add("toucan")
+        al.add("drake")
+        al.add("red-breasted merganser, Mergus serrator")
+        al.add("goose")
+        al.add("black swan, Cygnus atratus")
+        al.add("tusker")
+        al.add("echidna, spiny anteater, anteater")
+        al.add("platypus, duckbill, duckbilled platypus, duck-billed platypus, Ornithorhynchus anatinus")
+        al.add("wallaby, brush kangaroo")
+        al.add("koala, koala bear, kangaroo bear, native bear, Phascolarctos cinereus")
+        al.add("wombat")
+        al.add("jellyfish")
+        al.add("sea anemone, anemone")
+        al.add("brain coral")
+        al.add("flatworm, platyhelminth")
+        al.add("nematode, nematode worm, roundworm")
+        al.add("conch")
+        al.add("snail")
+        al.add("slug")
+        al.add("sea slug, nudibranch")
+        al.add("chiton, coat-of-mail shell, sea cradle, polyplacophore")
+        al.add("chambered nautilus, pearly nautilus, nautilus")
+        al.add("Dungeness crab, Cancer magister")
+        al.add("rock crab, Cancer irroratus")
+        al.add("fiddler crab")
+        al.add("king crab, Alaska crab, Alaskan king crab, Alaska king crab, Paralithodes camtschatica")
+        al.add("American lobster, Northern lobster, Maine lobster, Homarus americanus")
+        al.add("spiny lobster, langouste, rock lobster, crawfish, crayfish, sea crawfish")
+        al.add("crayfish, crawfish, crawdad, crawdaddy")
+        al.add("hermit crab")
+        al.add("isopod")
+        al.add("white stork, Ciconia ciconia")
+        al.add("black stork, Ciconia nigra")
+        al.add("spoonbill")
+        al.add("flamingo")
+        al.add("little blue heron, Egretta caerulea")
+        al.add("American egret, great white heron, Egretta albus")
+        al.add("bittern")
+        al.add("crane")
+        al.add("limpkin, Aramus pictus")
+        al.add("European gallinule, Porphyrio porphyrio")
+        al.add("American coot, marsh hen, mud hen, water hen, Fulica americana")
+        al.add("bustard")
+        al.add("ruddy turnstone, Arenaria interpres")
+        al.add("red-backed sandpiper, dunlin, Erolia alpina")
+        al.add("redshank, Tringa totanus")
+        al.add("dowitcher")
+        al.add("oystercatcher, oyster catcher")
+        al.add("pelican")
+        al.add("king penguin, Aptenodytes patagonica")
+        al.add("albatross, mollymawk")
+        al.add("grey whale, gray whale, devilfish, Eschrichtius gibbosus, Eschrichtius robustus")
+        al.add("killer whale, killer, orca, grampus, sea wolf, Orcinus orca")
+        al.add("dugong, Dugong dugon")
+        al.add("sea lion")
+        al.add("Chihuahua")
+        al.add("Japanese spaniel")
+        al.add("Maltese dog, Maltese terrier, Maltese")
+        al.add("Pekinese, Pekingese, Peke")
+        al.add("Shih-Tzu")
+        al.add("Blenheim spaniel")
+        al.add("papillon")
+        al.add("toy terrier")
+        al.add("Rhodesian ridgeback")
+        al.add("Afghan hound, Afghan")
+        al.add("basset, basset hound")
+        al.add("beagle")
+        al.add("bloodhound, sleuthhound")
+        al.add("bluetick")
+        al.add("black-and-tan coonhound")
+        al.add("Walker hound, Walker foxhound")
+        al.add("English foxhound")
+        al.add("redbone")
+        al.add("borzoi, Russian wolfhound")
+        al.add("Irish wolfhound")
+        al.add("Italian greyhound")
+        al.add("whippet")
+        al.add("Ibizan hound, Ibizan Podenco")
+        al.add("Norwegian elkhound, elkhound")
+        al.add("otterhound, otter hound")
+        al.add("Saluki, gazelle hound")
+        al.add("Scottish deerhound, deerhound")
+        al.add("Weimaraner")
+        al.add("Staffordshire bullterrier, Staffordshire bull terrier")
+        al.add("American Staffordshire terrier, Staffordshire terrier, American pit bull terrier, pit bull terrier")
+        al.add("Bedlington terrier")
+        al.add("Border terrier")
+        al.add("Kerry blue terrier")
+        al.add("Irish terrier")
+        al.add("Norfolk terrier")
+        al.add("Norwich terrier")
+        al.add("Yorkshire terrier")
+        al.add("wire-haired fox terrier")
+        al.add("Lakeland terrier")
+        al.add("Sealyham terrier, Sealyham")
+        al.add("Airedale, Airedale terrier")
+        al.add("cairn, cairn terrier")
+        al.add("Australian terrier")
+        al.add("Dandie Dinmont, Dandie Dinmont terrier")
+        al.add("Boston bull, Boston terrier")
+        al.add("miniature schnauzer")
+        al.add("giant schnauzer")
+        al.add("standard schnauzer")
+        al.add("Scotch terrier, Scottish terrier, Scottie")
+        al.add("Tibetan terrier, chrysanthemum dog")
+        al.add("silky terrier, Sydney silky")
+        al.add("soft-coated wheaten terrier")
+        al.add("West Highland white terrier")
+        al.add("Lhasa, Lhasa apso")
+        al.add("flat-coated retriever")
+        al.add("curly-coated retriever")
+        al.add("golden retriever")
+        al.add("Labrador retriever")
+        al.add("Chesapeake Bay retriever")
+        al.add("German short-haired pointer")
+        al.add("vizsla, Hungarian pointer")
+        al.add("English setter")
+        al.add("Irish setter, red setter")
+        al.add("Gordon setter")
+        al.add("Brittany spaniel")
+        al.add("clumber, clumber spaniel")
+        al.add("English springer, English springer spaniel")
+        al.add("Welsh springer spaniel")
+        al.add("cocker spaniel, English cocker spaniel, cocker")
+        al.add("Sussex spaniel")
+        al.add("Irish water spaniel")
+        al.add("kuvasz")
+        al.add("schipperke")
+        al.add("groenendael")
+        al.add("malinois")
+        al.add("briard")
+        al.add("kelpie")
+        al.add("komondor")
+        al.add("Old English sheepdog, bobtail")
+        al.add("Shetland sheepdog, Shetland sheep dog, Shetland")
+        al.add("collie")
+        al.add("Border collie")
+        al.add("Bouvier des Flandres, Bouviers des Flandres")
+        al.add("Rottweiler")
+        al.add("German shepherd, German shepherd dog, German police dog, alsatian")
+        al.add("Doberman, Doberman pinscher")
+        al.add("miniature pinscher")
+        al.add("Greater Swiss Mountain dog")
+        al.add("Bernese mountain dog")
+        al.add("Appenzeller")
+        al.add("EntleBucher")
+        al.add("boxer")
+        al.add("bull mastiff")
+        al.add("Tibetan mastiff")
+        al.add("French bulldog")
+        al.add("Great Dane")
+        al.add("Saint Bernard, St Bernard")
+        al.add("Eskimo dog, husky")
+        al.add("malamute, malemute, Alaskan malamute")
+        al.add("Siberian husky")
+        al.add("dalmatian, coach dog, carriage dog")
+        al.add("affenpinscher, monkey pinscher, monkey dog")
+        al.add("basenji")
+        al.add("pug, pug-dog")
+        al.add("Leonberg")
+        al.add("Newfoundland, Newfoundland dog")
+        al.add("Great Pyrenees")
+        al.add("Samoyed, Samoyede")
+        al.add("Pomeranian")
+        al.add("chow, chow chow")
+        al.add("keeshond")
+        al.add("Brabancon griffon")
+        al.add("Pembroke, Pembroke Welsh corgi")
+        al.add("Cardigan, Cardigan Welsh corgi")
+        al.add("toy poodle")
+        al.add("miniature poodle")
+        al.add("standard poodle")
+        al.add("Mexican hairless")
+        al.add("timber wolf, grey wolf, gray wolf, Canis lupus")
+        al.add("white wolf, Arctic wolf, Canis lupus tundrarum")
+        al.add("red wolf, maned wolf, Canis rufus, Canis niger")
+        al.add("coyote, prairie wolf, brush wolf, Canis latrans")
+        al.add("dingo, warrigal, warragal, Canis dingo")
+        al.add("dhole, Cuon alpinus")
+        al.add("African hunting dog, hyena dog, Cape hunting dog, Lycaon pictus")
+        al.add("hyena, hyaena")
+        al.add("red fox, Vulpes vulpes")
+        al.add("kit fox, Vulpes macrotis")
+        al.add("Arctic fox, white fox, Alopex lagopus")
+        al.add("grey fox, gray fox, Urocyon cinereoargenteus")
+        al.add("tabby, tabby cat")
+        al.add("tiger cat")
+        al.add("Persian cat")
+        al.add("Siamese cat, Siamese")
+        al.add("Egyptian cat")
+        al.add("cougar, puma, catamount, mountain lion, painter, panther, Felis concolor")
+        al.add("lynx, catamount")
+        al.add("leopard, Panthera pardus")
+        al.add("snow leopard, ounce, Panthera uncia")
+        al.add("jaguar, panther, Panthera onca, Felis onca")
+        al.add("lion, king of beasts, Panthera leo")
+        al.add("tiger, Panthera tigris")
+        al.add("cheetah, chetah, Acinonyx jubatus")
+        al.add("brown bear, bruin, Ursus arctos")
+        al.add("American black bear, black bear, Ursus americanus, Euarctos americanus")
+        al.add("ice bear, polar bear, Ursus Maritimus, Thalarctos maritimus")
+        al.add("sloth bear, Melursus ursinus, Ursus ursinus")
+        al.add("mongoose")
+        al.add("meerkat, mierkat")
+        al.add("tiger beetle")
+        al.add("ladybug, ladybeetle, lady beetle, ladybird, ladybird beetle")
+        al.add("ground beetle, carabid beetle")
+        al.add("long-horned beetle, longicorn, longicorn beetle")
+        al.add("leaf beetle, chrysomelid")
+        al.add("dung beetle")
+        al.add("rhinoceros beetle")
+        al.add("weevil")
+        al.add("fly")
+        al.add("bee")
+        al.add("ant, emmet, pismire")
+        al.add("grasshopper, hopper")
+        al.add("cricket")
+        al.add("walking stick, walkingstick, stick insect")
+        al.add("cockroach, roach")
+        al.add("mantis, mantid")
+        al.add("cicada, cicala")
+        al.add("leafhopper")
+        al.add("lacewing, lacewing fly")
+        al.add("dragonfly, darning needle, devil''s darning needle, sewing needle, snake feeder, snake doctor, mosquito hawk, skeeter hawk")
+        al.add("damselfly")
+        al.add("admiral")
+        al.add("ringlet, ringlet butterfly")
+        al.add("monarch, monarch butterfly, milkweed butterfly, Danaus plexippus")
+        al.add("cabbage butterfly")
+        al.add("sulphur butterfly, sulfur butterfly")
+        al.add("lycaenid, lycaenid butterfly")
+        al.add("starfish, sea star")
+        al.add("sea urchin")
+        al.add("sea cucumber, holothurian")
+        al.add("wood rabbit, cottontail, cottontail rabbit")
+        al.add("hare")
+        al.add("Angora, Angora rabbit")
+        al.add("hamster")
+        al.add("porcupine, hedgehog")
+        al.add("fox squirrel, eastern fox squirrel, Sciurus niger")
+        al.add("marmot")
+        al.add("beaver")
+        al.add("guinea pig, Cavia cobaya")
+        al.add("sorrel")
+        al.add("zebra")
+        al.add("hog, pig, grunter, squealer, Sus scrofa")
+        al.add("wild boar, boar, Sus scrofa")
+        al.add("warthog")
+        al.add("hippopotamus, hippo, river horse, Hippopotamus amphibius")
+        al.add("ox")
+        al.add("water buffalo, water ox, Asiatic buffalo, Bubalus bubalis")
+        al.add("bison")
+        al.add("ram, tup")
+        al.add("bighorn, bighorn sheep, cimarron, Rocky Mountain bighorn, Rocky Mountain sheep, Ovis canadensis")
+        al.add("ibex, Capra ibex")
+        al.add("hartebeest")
+        al.add("impala, Aepyceros melampus")
+        al.add("gazelle")
+        al.add("Arabian camel, dromedary, Camelus dromedarius")
+        al.add("llama")
+        al.add("weasel")
+        al.add("mink")
+        al.add("polecat, fitch, foulmart, foumart, Mustela putorius")
+        al.add("black-footed ferret, ferret, Mustela nigripes")
+        al.add("otter")
+        al.add("skunk, polecat, wood pussy")
+        al.add("badger")
+        al.add("armadillo")
+        al.add("three-toed sloth, ai, Bradypus tridactylus")
+        al.add("orangutan, orang, orangutang, Pongo pygmaeus")
+        al.add("gorilla, Gorilla gorilla")
+        al.add("chimpanzee, chimp, Pan troglodytes")
+        al.add("gibbon, Hylobates lar")
+        al.add("siamang, Hylobates syndactylus, Symphalangus syndactylus")
+        al.add("guenon, guenon monkey")
+        al.add("patas, hussar monkey, Erythrocebus patas")
+        al.add("baboon")
+        al.add("macaque")
+        al.add("langur")
+        al.add("colobus, colobus monkey")
+        al.add("proboscis monkey, Nasalis larvatus")
+        al.add("marmoset")
+        al.add("capuchin, ringtail, Cebus capucinus")
+        al.add("howler monkey, howler")
+        al.add("titi, titi monkey")
+        al.add("spider monkey, Ateles geoffroyi")
+        al.add("squirrel monkey, Saimiri sciureus")
+        al.add("Madagascar cat, ring-tailed lemur, Lemur catta")
+        al.add("indri, indris, Indri indri, Indri brevicaudatus")
+        al.add("Indian elephant, Elephas maximus")
+        al.add("African elephant, Loxodonta africana")
+        al.add("lesser panda, red panda, panda, bear cat, cat bear, Ailurus fulgens")
+        al.add("giant panda, panda, panda bear, coon bear, Ailuropoda melanoleuca")
+        al.add("barracouta, snoek")
+        al.add("eel")
+        al.add("coho, cohoe, coho salmon, blue jack, silver salmon, Oncorhynchus kisutch")
+        al.add("rock beauty, Holocanthus tricolor")
+        al.add("anemone fish")
+        al.add("sturgeon")
+        al.add("gar, garfish, garpike, billfish, Lepisosteus osseus")
+        al.add("lionfish")
+        al.add("puffer, pufferfish, blowfish, globefish")
+        al.add("abacus")
+        al.add("abaya")
+        al.add("academic gown, academic robe, judge''s robe")
+        al.add("accordion, piano accordion, squeeze box")
+        al.add("acoustic guitar")
+        al.add("aircraft carrier, carrier, flattop, attack aircraft carrier")
+        al.add("airliner")
+        al.add("airship, dirigible")
+        al.add("altar")
+        al.add("ambulance")
+        al.add("amphibian, amphibious vehicle")
+        al.add("analog clock")
+        al.add("apiary, bee house")
+        al.add("apron")
+        al.add("ashcan, trash can, garbage can, wastebin, ash bin, ash-bin, ashbin, dustbin, trash barrel, trash bin")
+        al.add("assault rifle, assault gun")
+        al.add("backpack, back pack, knapsack, packsack, rucksack, haversack")
+        al.add("bakery, bakeshop, bakehouse")
+        al.add("balance beam, beam")
+        al.add("balloon")
+        al.add("ballpoint, ballpoint pen, ballpen, Biro")
+        al.add("Band Aid")
+        al.add("banjo")
+        al.add("bannister, banister, balustrade, balusters, handrail")
+        al.add("barbell")
+        al.add("barber chair")
+        al.add("barbershop")
+        al.add("barn")
+        al.add("barometer")
+        al.add("barrel, cask")
+        al.add("barrow, garden cart, lawn cart, wheelbarrow")
+        al.add("baseball")
+        al.add("basketball")
+        al.add("bassinet")
+        al.add("bassoon")
+        al.add("bathing cap, swimming cap")
+        al.add("bath towel")
+        al.add("bathtub, bathing tub, bath, tub")
+        al.add("beach wagon, station wagon, wagon, estate car, beach waggon, station waggon, waggon")
+        al.add("beacon, lighthouse, beacon light, pharos")
+        al.add("beaker")
+        al.add("bearskin, busby, shako")
+        al.add("beer bottle")
+        al.add("beer glass")
+        al.add("bell cote, bell cot")
+        al.add("bib")
+        al.add("bicycle-built-for-two, tandem bicycle, tandem")
+        al.add("bikini, two-piece")
+        al.add("binder, ring-binder")
+        al.add("binoculars, field glasses, opera glasses")
+        al.add("birdhouse")
+        al.add("boathouse")
+        al.add("bobsled, bobsleigh, bob")
+        al.add("bolo tie, bolo, bola tie, bola")
+        al.add("bonnet, poke bonnet")
+        al.add("bookcase")
+        al.add("bookshop, bookstore, bookstall")
+        al.add("bottlecap")
+        al.add("bow")
+        al.add("bow tie, bow-tie, bowtie")
+        al.add("brass, memorial tablet, plaque")
+        al.add("brassiere, bra, bandeau")
+        al.add("breakwater, groin, groyne, mole, bulwark, seawall, jetty")
+        al.add("breastplate, aegis, egis")
+        al.add("broom")
+        al.add("bucket, pail")
+        al.add("buckle")
+        al.add("bulletproof vest")
+        al.add("bullet train, bullet")
+        al.add("butcher shop, meat market")
+        al.add("cab, hack, taxi, taxicab")
+        al.add("caldron, cauldron")
+        al.add("candle, taper, wax light")
+        al.add("cannon")
+        al.add("canoe")
+        al.add("can opener, tin opener")
+        al.add("cardigan")
+        al.add("car mirror")
+        al.add("carousel, carrousel, merry-go-round, roundabout, whirligig")
+        al.add("carpenter''s kit, tool kit")
+        al.add("carton")
+        al.add("car wheel")
+        al.add("cash machine, cash dispenser, automated teller machine, automatic teller machine, automated teller, automatic teller, ATM")
+        al.add("cassette")
+        al.add("cassette player")
+        al.add("castle")
+        al.add("catamaran")
+        al.add("CD player")
+        al.add("cello, violoncello")
+        al.add("cellular telephone, cellular phone, cellphone, cell, mobile phone")
+        al.add("chain")
+        al.add("chainlink fence")
+        al.add("chain mail, ring mail, mail, chain armor, chain armour, ring armor, ring armour")
+        al.add("chain saw, chainsaw")
+        al.add("chest")
+        al.add("chiffonier, commode")
+        al.add("chime, bell, gong")
+        al.add("china cabinet, china closet")
+        al.add("Christmas stocking")
+        al.add("church, church building")
+        al.add("cinema, movie theater, movie theatre, movie house, picture palace")
+        al.add("cleaver, meat cleaver, chopper")
+        al.add("cliff dwelling")
+        al.add("cloak")
+        al.add("clog, geta, patten, sabot")
+        al.add("cocktail shaker")
+        al.add("coffee mug")
+        al.add("coffeepot")
+        al.add("coil, spiral, volute, whorl, helix")
+        al.add("combination lock")
+        al.add("computer keyboard, keypad")
+        al.add("confectionery, confectionary, candy store")
+        al.add("container ship, containership, container vessel")
+        al.add("convertible")
+        al.add("corkscrew, bottle screw")
+        al.add("cornet, horn, trumpet, trump")
+        al.add("cowboy boot")
+        al.add("cowboy hat, ten-gallon hat")
+        al.add("cradle")
+        al.add("crane")
+        al.add("crash helmet")
+        al.add("crate")
+        al.add("crib, cot")
+        al.add("Crock Pot")
+        al.add("croquet ball")
+        al.add("crutch")
+        al.add("cuirass")
+        al.add("dam, dike, dyke")
+        al.add("desk")
+        al.add("desktop computer")
+        al.add("dial telephone, dial phone")
+        al.add("diaper, nappy, napkin")
+        al.add("digital clock")
+        al.add("digital watch")
+        al.add("dining table, board")
+        al.add("dishrag, dishcloth")
+        al.add("dishwasher, dish washer, dishwashing machine")
+        al.add("disk brake, disc brake")
+        al.add("dock, dockage, docking facility")
+        al.add("dogsled, dog sled, dog sleigh")
+        al.add("dome")
+        al.add("doormat, welcome mat")
+        al.add("drilling platform, offshore rig")
+        al.add("drum, membranophone, tympan")
+        al.add("drumstick")
+        al.add("dumbbell")
+        al.add("Dutch oven")
+        al.add("electric fan, blower")
+        al.add("electric guitar")
+        al.add("electric locomotive")
+        al.add("entertainment center")
+        al.add("envelope")
+        al.add("espresso maker")
+        al.add("face powder")
+        al.add("feather boa, boa")
+        al.add("file, file cabinet, filing cabinet")
+        al.add("fireboat")
+        al.add("fire engine, fire truck")
+        al.add("fire screen, fireguard")
+        al.add("flagpole, flagstaff")
+        al.add("flute, transverse flute")
+        al.add("folding chair")
+        al.add("football helmet")
+        al.add("forklift")
+        al.add("fountain")
+        al.add("fountain pen")
+        al.add("four-poster")
+        al.add("freight car")
+        al.add("French horn, horn")
+        al.add("frying pan, frypan, skillet")
+        al.add("fur coat")
+        al.add("garbage truck, dustcart")
+        al.add("gasmask, respirator, gas helmet")
+        al.add("gas pump, gasoline pump, petrol pump, island dispenser")
+        al.add("goblet")
+        al.add("go-kart")
+        al.add("golf ball")
+        al.add("golfcart, golf cart")
+        al.add("gondola")
+        al.add("gong, tam-tam")
+        al.add("gown")
+        al.add("grand piano, grand")
+        al.add("greenhouse, nursery, glasshouse")
+        al.add("grille, radiator grille")
+        al.add("grocery store, grocery, food market, market")
+        al.add("guillotine")
+        al.add("hair slide")
+        al.add("hair spray")
+        al.add("half track")
+        al.add("hammer")
+        al.add("hamper")
+        al.add("hand blower, blow dryer, blow drier, hair dryer, hair drier")
+        al.add("hand-held computer, hand-held microcomputer")
+        al.add("handkerchief, hankie, hanky, hankey")
+        al.add("hard disc, hard disk, fixed disk")
+        al.add("harmonica, mouth organ, harp, mouth harp")
+        al.add("harp")
+        al.add("harvester, reaper")
+        al.add("hatchet")
+        al.add("holster")
+        al.add("home theater, home theatre")
+        al.add("honeycomb")
+        al.add("hook, claw")
+        al.add("hoopskirt, crinoline")
+        al.add("horizontal bar, high bar")
+        al.add("horse cart, horse-cart")
+        al.add("hourglass")
+        al.add("iPod")
+        al.add("iron, smoothing iron")
+        al.add("jack-o''-lantern")
+        al.add("jean, blue jean, denim")
+        al.add("jeep, landrover")
+        al.add("jersey, T-shirt, tee shirt")
+        al.add("jigsaw puzzle")
+        al.add("jinrikisha, ricksha, rickshaw")
+        al.add("joystick")
+        al.add("kimono")
+        al.add("knee pad")
+        al.add("knot")
+        al.add("lab coat, laboratory coat")
+        al.add("ladle")
+        al.add("lampshade, lamp shade")
+        al.add("laptop, laptop computer")
+        al.add("lawn mower, mower")
+        al.add("lens cap, lens cover")
+        al.add("letter opener, paper knife, paperknife")
+        al.add("library")
+        al.add("lifeboat")
+        al.add("lighter, light, igniter, ignitor")
+        al.add("limousine, limo")
+        al.add("liner, ocean liner")
+        al.add("lipstick, lip rouge")
+        al.add("Loafer")
+        al.add("lotion")
+        al.add("loudspeaker, speaker, speaker unit, loudspeaker system, speaker system")
+        al.add("loupe, jeweler''s loupe")
+        al.add("lumbermill, sawmill")
+        al.add("magnetic compass")
+        al.add("mailbag, postbag")
+        al.add("mailbox, letter box")
+        al.add("maillot")
+        al.add("maillot, tank suit")
+        al.add("manhole cover")
+        al.add("maraca")
+        al.add("marimba, xylophone")
+        al.add("mask")
+        al.add("matchstick")
+        al.add("maypole")
+        al.add("maze, labyrinth")
+        al.add("measuring cup")
+        al.add("medicine chest, medicine cabinet")
+        al.add("megalith, megalithic structure")
+        al.add("microphone, mike")
+        al.add("microwave, microwave oven")
+        al.add("military uniform")
+        al.add("milk can")
+        al.add("minibus")
+        al.add("miniskirt, mini")
+        al.add("minivan")
+        al.add("missile")
+        al.add("mitten")
+        al.add("mixing bowl")
+        al.add("mobile home, manufactured home")
+        al.add("Model T")
+        al.add("modem")
+        al.add("monastery")
+        al.add("monitor")
+        al.add("moped")
+        al.add("mortar")
+        al.add("mortarboard")
+        al.add("mosque")
+        al.add("mosquito net")
+        al.add("motor scooter, scooter")
+        al.add("mountain bike, all-terrain bike, off-roader")
+        al.add("mountain tent")
+        al.add("mouse, computer mouse")
+        al.add("mousetrap")
+        al.add("moving van")
+        al.add("muzzle")
+        al.add("nail")
+        al.add("neck brace")
+        al.add("necklace")
+        al.add("nipple")
+        al.add("notebook, notebook computer")
+        al.add("obelisk")
+        al.add("oboe, hautboy, hautbois")
+        al.add("ocarina, sweet potato")
+        al.add("odometer, hodometer, mileometer, milometer")
+        al.add("oil filter")
+        al.add("organ, pipe organ")
+        al.add("oscilloscope, scope, cathode-ray oscilloscope, CRO")
+        al.add("overskirt")
+        al.add("oxcart")
+        al.add("oxygen mask")
+        al.add("packet")
+        al.add("paddle, boat paddle")
+        al.add("paddlewheel, paddle wheel")
+        al.add("padlock")
+        al.add("paintbrush")
+        al.add("pajama, pyjama, pj''s, jammies")
+        al.add("palace")
+        al.add("panpipe, pandean pipe, syrinx")
+        al.add("paper towel")
+        al.add("parachute, chute")
+        al.add("parallel bars, bars")
+        al.add("park bench")
+        al.add("parking meter")
+        al.add("passenger car, coach, carriage")
+        al.add("patio, terrace")
+        al.add("pay-phone, pay-station")
+        al.add("pedestal, plinth, footstall")
+        al.add("pencil box, pencil case")
+        al.add("pencil sharpener")
+        al.add("perfume, essence")
+        al.add("Petri dish")
+        al.add("photocopier")
+        al.add("pick, plectrum, plectron")
+        al.add("pickelhaube")
+        al.add("picket fence, paling")
+        al.add("pickup, pickup truck")
+        al.add("pier")
+        al.add("piggy bank, penny bank")
+        al.add("pill bottle")
+        al.add("pillow")
+        al.add("ping-pong ball")
+        al.add("pinwheel")
+        al.add("pirate, pirate ship")
+        al.add("pitcher, ewer")
+        al.add("plane, carpenter''s plane, woodworking plane")
+        al.add("planetarium")
+        al.add("plastic bag")
+        al.add("plate rack")
+        al.add("plow, plough")
+        al.add("plunger, plumber''s helper")
+        al.add("Polaroid camera, Polaroid Land camera")
+        al.add("pole")
+        al.add("police van, police wagon, paddy wagon, patrol wagon, wagon, black Maria")
+        al.add("poncho")
+        al.add("pool table, billiard table, snooker table")
+        al.add("pop bottle, soda bottle")
+        al.add("pot, flowerpot")
+        al.add("potter''s wheel")
+        al.add("power drill")
+        al.add("prayer rug, prayer mat")
+        al.add("printer")
+        al.add("prison, prison house")
+        al.add("projectile, missile")
+        al.add("projector")
+        al.add("puck, hockey puck")
+        al.add("punching bag, punch bag, punching ball, punchball")
+        al.add("purse")
+        al.add("quill, quill pen")
+        al.add("quilt, comforter, comfort, puff")
+        al.add("racer, race car, racing car")
+        al.add("racket, racquet")
+        al.add("radiator")
+        al.add("radio, wireless")
+        al.add("radio telescope, radio reflector")
+        al.add("rain barrel")
+        al.add("recreational vehicle, RV, R.V.")
+        al.add("reel")
+        al.add("reflex camera")
+        al.add("refrigerator, icebox")
+        al.add("remote control, remote")
+        al.add("restaurant, eating house, eating place, eatery")
+        al.add("revolver, six-gun, six-shooter")
+        al.add("rifle")
+        al.add("rocking chair, rocker")
+        al.add("rotisserie")
+        al.add("rubber eraser, rubber, pencil eraser")
+        al.add("rugby ball")
+        al.add("rule, ruler")
+        al.add("running shoe")
+        al.add("safe")
+        al.add("safety pin")
+        al.add("saltshaker, salt shaker")
+        al.add("sandal")
+        al.add("sarong")
+        al.add("sax, saxophone")
+        al.add("scabbard")
+        al.add("scale, weighing machine")
+        al.add("school bus")
+        al.add("schooner")
+        al.add("scoreboard")
+        al.add("screen, CRT screen")
+        al.add("screw")
+        al.add("screwdriver")
+        al.add("seat belt, seatbelt")
+        al.add("sewing machine")
+        al.add("shield, buckler")
+        al.add("shoe shop, shoe-shop, shoe store")
+        al.add("shoji")
+        al.add("shopping basket")
+        al.add("shopping cart")
+        al.add("shovel")
+        al.add("shower cap")
+        al.add("shower curtain")
+        al.add("ski")
+        al.add("ski mask")
+        al.add("sleeping bag")
+        al.add("slide rule, slipstick")
+        al.add("sliding door")
+        al.add("slot, one-armed bandit")
+        al.add("snorkel")
+        al.add("snowmobile")
+        al.add("snowplow, snowplough")
+        al.add("soap dispenser")
+        al.add("soccer ball")
+        al.add("sock")
+        al.add("solar dish, solar collector, solar furnace")
+        al.add("sombrero")
+        al.add("soup bowl")
+        al.add("space bar")
+        al.add("space heater")
+        al.add("space shuttle")
+        al.add("spatula")
+        al.add("speedboat")
+        al.add("spider web, spider''s web")
+        al.add("spindle")
+        al.add("sports car, sport car")
+        al.add("spotlight, spot")
+        al.add("stage")
+        al.add("steam locomotive")
+        al.add("steel arch bridge")
+        al.add("steel drum")
+        al.add("stethoscope")
+        al.add("stole")
+        al.add("stone wall")
+        al.add("stopwatch, stop watch")
+        al.add("stove")
+        al.add("strainer")
+        al.add("streetcar, tram, tramcar, trolley, trolley car")
+        al.add("stretcher")
+        al.add("studio couch, day bed")
+        al.add("stupa, tope")
+        al.add("submarine, pigboat, sub, U-boat")
+        al.add("suit, suit of clothes")
+        al.add("sundial")
+        al.add("sunglass")
+        al.add("sunglasses, dark glasses, shades")
+        al.add("sunscreen, sunblock, sun blocker")
+        al.add("suspension bridge")
+        al.add("swab, swob, mop")
+        al.add("sweatshirt")
+        al.add("swimming trunks, bathing trunks")
+        al.add("swing")
+        al.add("switch, electric switch, electrical switch")
+        al.add("syringe")
+        al.add("table lamp")
+        al.add("tank, army tank, armored combat vehicle, armoured combat vehicle")
+        al.add("tape player")
+        al.add("teapot")
+        al.add("teddy, teddy bear")
+        al.add("television, television system")
+        al.add("tennis ball")
+        al.add("thatch, thatched roof")
+        al.add("theater curtain, theatre curtain")
+        al.add("thimble")
+        al.add("thresher, thrasher, threshing machine")
+        al.add("throne")
+        al.add("tile roof")
+        al.add("toaster")
+        al.add("tobacco shop, tobacconist shop, tobacconist")
+        al.add("toilet seat")
+        al.add("torch")
+        al.add("totem pole")
+        al.add("tow truck, tow car, wrecker")
+        al.add("toyshop")
+        al.add("tractor")
+        al.add("trailer truck, tractor trailer, trucking rig, rig, articulated lorry, semi")
+        al.add("tray")
+        al.add("trench coat")
+        al.add("tricycle, trike, velocipede")
+        al.add("trimaran")
+        al.add("tripod")
+        al.add("triumphal arch")
+        al.add("trolleybus, trolley coach, trackless trolley")
+        al.add("trombone")
+        al.add("tub, vat")
+        al.add("turnstile")
+        al.add("typewriter keyboard")
+        al.add("umbrella")
+        al.add("unicycle, monocycle")
+        al.add("upright, upright piano")
+        al.add("vacuum, vacuum cleaner")
+        al.add("vase")
+        al.add("vault")
+        al.add("velvet")
+        al.add("vending machine")
+        al.add("vestment")
+        al.add("viaduct")
+        al.add("violin, fiddle")
+        al.add("volleyball")
+        al.add("waffle iron")
+        al.add("wall clock")
+        al.add("wallet, billfold, notecase, pocketbook")
+        al.add("wardrobe, closet, press")
+        al.add("warplane, military plane")
+        al.add("washbasin, handbasin, washbowl, lavabo, wash-hand basin")
+        al.add("washer, automatic washer, washing machine")
+        al.add("water bottle")
+        al.add("water jug")
+        al.add("water tower")
+        al.add("whiskey jug")
+        al.add("whistle")
+        al.add("wig")
+        al.add("window screen")
+        al.add("window shade")
+        al.add("Windsor tie")
+        al.add("wine bottle")
+        al.add("wing")
+        al.add("wok")
+        al.add("wooden spoon")
+        al.add("wool, woolen, woollen")
+        al.add("worm fence, snake fence, snake-rail fence, Virginia fence")
+        al.add("wreck")
+        al.add("yawl")
+        al.add("yurt")
+        al.add("web site, website, internet site, site")
+        al.add("comic book")
+        al.add("crossword puzzle, crossword")
+        al.add("street sign")
+        al.add("traffic light, traffic signal, stoplight")
+        al.add("book jacket, dust cover, dust jacket, dust wrapper")
+        al.add("menu")
+        al.add("plate")
+        al.add("guacamole")
+        al.add("consomme")
+        al.add("hot pot, hotpot")
+        al.add("trifle")
+        al.add("ice cream, icecream")
+        al.add("ice lolly, lolly, lollipop, popsicle")
+        al.add("French loaf")
+        al.add("bagel, beigel")
+        al.add("pretzel")
+        al.add("cheeseburger")
+        al.add("hotdog, hot dog, red hot")
+        al.add("mashed potato")
+        al.add("head cabbage")
+        al.add("broccoli")
+        al.add("cauliflower")
+        al.add("zucchini, courgette")
+        al.add("spaghetti squash")
+        al.add("acorn squash")
+        al.add("butternut squash")
+        al.add("cucumber, cuke")
+        al.add("artichoke, globe artichoke")
+        al.add("bell pepper")
+        al.add("cardoon")
+        al.add("mushroom")
+        al.add("Granny Smith")
+        al.add("strawberry")
+        al.add("orange")
+        al.add("lemon")
+        al.add("fig")
+        al.add("pineapple, ananas")
+        al.add("banana")
+        al.add("jackfruit, jak, jack")
+        al.add("custard apple")
+        al.add("pomegranate")
+        al.add("hay")
+        al.add("carbonara")
+        al.add("chocolate sauce, chocolate syrup")
+        al.add("dough")
+        al.add("meat loaf, meatloaf")
+        al.add("pizza, pizza pie")
+        al.add("potpie")
+        al.add("burrito")
+        al.add("red wine")
+        al.add("espresso")
+        al.add("cup")
+        al.add("eggnog")
+        al.add("alp")
+        al.add("bubble")
+        al.add("cliff, drop, drop-off")
+        al.add("coral reef")
+        al.add("geyser")
+        al.add("lakeside, lakeshore")
+        al.add("promontory, headland, head, foreland")
+        al.add("sandbar, sand bar")
+        al.add("seashore, coast, seacoast, sea-coast")
+        al.add("valley, vale")
+        al.add("volcano")
+        al.add("ballplayer, baseball player")
+        al.add("groom, bridegroom")
+        al.add("scuba diver")
+        al.add("rapeseed")
+        al.add("daisy")
+        al.add("yellow lady''s slipper, yellow lady-slipper, Cypripedium calceolus, Cypripedium parviflorum")
+        al.add("corn")
+        al.add("acorn")
+        al.add("hip, rose hip, rosehip")
+        al.add("buckeye, horse chestnut, conker")
+        al.add("coral fungus")
+        al.add("agaric")
+        al.add("gyromitra")
+        al.add("stinkhorn, carrion fungus")
+        al.add("earthstar")
+        al.add("hen-of-the-woods, hen of the woods, Polyporus frondosus, Grifola frondosa")
+        al.add("bolete")
+        al.add("ear, spike, capitulum")
+        al.add("toilet tissue, toilet paper, bathroom tissue")
         al
     }
 
